\name{accessors}
\alias{NetAffxResource-accessors}

\alias{annotationsFile}
\alias{directory}
\alias{password}
\alias{user}

\docType{data}
\title{Access NetAffxResource-class content}
\description{

  These functions exist to access the information represented in objects
  of \code{linkS4class{NetAffxResource}} and related classes.

}
\usage{
annotationsFile(x)
directory(x)
password(x)
user(x)
}
\arguments{
  \item{x}{An object for which the accessor is defined.}
}
\value{

  See, e.g., \code{help('directory,NetAffxResource-method')} for values
  returned by specific methods.

}
\seealso{\code{\linkS4class{NetAffxResource}}}
\author{Martin Morgan}
\keyword{methods}
