\name{AffyCompatible-package}
\alias{AffyCompatible-package}
\alias{AffyCompatible}
\docType{package}
\title{

  Affymetrix chip annotation and sample attribute management

}
\description{

  This package manages Affymetrix NetAffx annotation files, and parses
  MAGE and ARR sample attribute files into R data structures.

  NetAffx provides a way to manage the diverse annotation files provided
  by Affymetrix for their arrays.

  ARR files provide attribute inforamtion about samples used in an
  experiment. These files are created, for instance, by the GeneChip
  Command Console or compliant applications, or

  MAGE files also provide information about samples. Support in this
  package is for MAGE produced by Affymetrix GCOS compliant
  applications, and emphasizes sample description and processing.

}
\details{

  Package details are available in the DESCRIPTION file of the package,
  or with the R command \code{packageDescription("AffyCompatible")}

  Overviews of how to use the functionality of this package are available
  in the package vignettes. See \code{vignette(package="AffyCompatible")}.

}
\author{

  Written by Martin Morgan.

  Maintainer: Martin Morgan <martin.t.morgan@gmail.com>

}
\keyword{ package }
