\name{zzz-ARRCompatibility-class}
\docType{class}
\alias{zzz-ARRCompatibility-class}

% generic accessors
\alias{affyBarcode}
\alias{arrayAttribute}
\alias{arrayName}
\alias{control}
\alias{createdBy}
\alias{createdDateTime}
\alias{createdStep}
\alias{defaultValue}
\alias{guid}
\alias{libraryPackageName}
\alias{masterFileGUID}
\alias{masterFileName}
\alias{mediaCol}
\alias{mediaFileGUID}
\alias{mediaFileName}
\alias{mediaRow}
\alias{mediaType}
\alias{namespace}
\alias{originalProjectName}
\alias{patassignmentMethod}
\alias{physicalArray}
\alias{physicalArrays}
\alias{required}
\alias{userAttribute}
\alias{userAttributes}
\alias{userAttributeValue}
\alias{value}
\alias{version}

% accessor methods
\alias{affxComment,PhysicalArray-method}
\alias{affyBarcode,PhysicalArray-method}
\alias{arrayAttribute,PhysicalArray-method}
\alias{arrayName,PhysicalArray-method}
\alias{control,UserAttribute-method}
\alias{createdBy,ArraySetFile-method}
\alias{createdBy,PhysicalArray-method}
\alias{createdBy,TemplateFile-method}
\alias{createdDateTime,ArraySetFile-method}
\alias{createdDateTime,PhysicalArray-method}
\alias{createdDateTime,TemplateFile-method}
\alias{createdStep,ArraySetFile-method}
\alias{createdStep,PhysicalArray-method}
\alias{defaultValue,UserAttribute-method}
\alias{guid,ArraySetFile-method}
\alias{guid,PhysicalArray-method}
\alias{guid,TemplateFile-method}
\alias{libraryPackageName,PhysicalArray-method}
\alias{masterFileGUID,PhysicalArray-method}
\alias{masterFileName,PhysicalArray-method}
\alias{mediaCol,PhysicalArray-method}
\alias{mediaFileGUID,PhysicalArray-method}
\alias{mediaFileName,PhysicalArray-method}
\alias{mediaRow,PhysicalArray-method}
\alias{mediaType,PhysicalArray-method}
\alias{name,ArrayAttribute-method}
\alias{name,UserAttribute-method}
\alias{namespace,UserAttribute-method}
\alias{originalProjectName,ArraySetFile-method}
\alias{patassignmentMethod,PhysicalArray-method}
\alias{physicalArray,PhysicalArrays-method}
\alias{physicalArrays,ArraySetFile-method}
\alias{required,UserAttribute-method}
\alias{type,ArraySetFile-method}
\alias{type,PhysicalArray-method}
\alias{type,TemplateFile-method}
\alias{type,UserAttribute-method}
\alias{userAttribute,UserAttributes-method}
\alias{userAttributes,ArraySetFile-method}
\alias{userAttributes,TemplateFile-method}
\alias{userAttributeValue,UserAttribute-method}
\alias{value,Control-method}
\alias{version,ArraySetFile-method}
\alias{version,TemplateFile-method}

\title{

  (Auto-generated) classes and methods for navigating ARR sample
  attributes.

}
\description{

  These classes and methods facilitate navigation of the Affymetrix ARR
  sample attribute files. ARR files are generated by the Affymetrix AGCC
  (GeneChip\textregistered{} Command Console) application. Specifically,
  the classes are R representations of the Affymetrix classes defined in
  ArraySetAndTemplateFile.dtd. The 'show' methods for these classs
  indicate how objects are to be accessed; details are in the
  `Retrieving MAGE and ARR sample attributes' vignette.

}
\section{Objects from the Class}{

  Objects are created internally, by parsing XML retrieved from ARR, and
  are not created directly by the user.

}
\section{Slots}{

  Objects have a variety of slots. To discover slot names and content
  type, see the \code{slotNames} portion of the example, below.

}
\section{Extends}{
Class \code{"\linkS4class{ARRCompatibility}"}, directly.
Class \code{"\linkS4class{AffyCompatible}"}, by class "ARRCompatibility", distance 2.
}
\section{Methods}{

  'Accessor' methods are defined on auto-generated classes to access slot
  content.

  Accessors are apparent from displaying the object; each line of the
  display is of the form \code{accessor: content}. To retrieve the
  content of an instance \code{object}, invoke \code{accessor(object)}.

  Some objects contains several instances of another object. These
  'typed associations' are represented by the
  \code{\linkS4class{.TypedAssociation}} class, and are navigated or
  subset as one would a standard R vector, e.g., using \code{[},
  \code{[[}. A useful paradigm for displaying all instances is
  illustrated in the vignette.

}
\author{Martin Morgan <mtmorgan@fhcrc.org>}
\seealso{

  Vignette `Retrieving MAGE and ARR sample attributes'.

}

\examples{
## defined and auto-generated classes
names(slot(getClass("ARRCompatibility",
                    where=getNamespace("AffyCompatible")),
           "subclasses"))


}

\keyword{classes}
