\name{zzz-DTTCompatibility-class}
\docType{class}
\alias{zzz-DTTCompatibility-class}

% generic accessors
\alias{accession}
\alias{accessions_assnlist}
\alias{accessionVersion}
\alias{action}
\alias{action_assn}
\alias{actionMeasurement_assn}
\alias{activityDate}
\alias{address}
\alias{adjustments_assnlist}
\alias{affiliation_assnref}
\alias{affxCategory}
\alias{affxEnd}
\alias{affxImage}
\alias{affxOrder}
\alias{affxPerson}
\alias{affxRow}
\alias{affxSequence}
\alias{affxStart}
\alias{affxText}
\alias{affxTitle}
\alias{analysisResults_assnreflist}
\alias{annotations_assnlist}
\alias{array_assnlist}
\alias{array_assnref}
\alias{arrayDesign}
\alias{arrayDesign_assnlist}
\alias{arrayDesign_assnref}
\alias{arrayDesign_package}
\alias{arrayDesign_ref}
\alias{arrayGroup}
\alias{arrayGroup_assnlist}
\alias{arrayGroup_assnref}
\alias{arrayGroup_ref}
\alias{arrayIdentifier}
\alias{arrayManufacture}
\alias{arrayManufacture_assnlist}
\alias{arrayManufactureDeviation}
\alias{arrayManufactureDeviations_assnlist}
\alias{arrayManufacture_ref}
\alias{arrayManufacturers_assnreflist}
\alias{array_package}
\alias{array_ref}
\alias{arrays_assnreflist}
\alias{arraySpacingX}
\alias{arraySpacingY}
\alias{arrayXOrigin}
\alias{arrayYOrigin}
\alias{associations_assnlist}
\alias{audit}
\alias{auditAndSecurity_package}
\alias{auditTrail_assnlist}
\alias{authors}
\alias{barcode}
\alias{basis}
\alias{bibliographicReference}
\alias{bibliographicReferences_assnlist}
\alias{bioAssay_assnlist}
\alias{bioAssay_assnref}
\alias{bioAssayCreation}
\alias{bioAssayCreation_assn}
\alias{bioAssayData_assnlist}
\alias{bioAssayData_assnreflist}
\alias{bioAssayDataCluster}
\alias{bioAssayDataCluster_assnlist}
\alias{bioAssayDataCluster_ref}
\alias{bioAssayData_package}
\alias{bioAssayDataSources_assnreflist}
\alias{bioAssayDimension}
\alias{bioAssayDimension_assnlist}
\alias{bioAssayDimension_assnref}
\alias{bioAssayDimension_ref}
\alias{bioAssayFactorValues_assnreflist}
\alias{bioAssayMap}
\alias{bioAssayMap_assnlist}
\alias{bioAssayMapping}
\alias{bioAssayMapping_assn}
\alias{bioAssayMap_ref}
\alias{bioAssayMaps_assnreflist}
\alias{bioAssayMapTarget_assnref}
\alias{bioAssay_package}
\alias{bioAssays_assnreflist}
\alias{bioAssayTreatment}
\alias{bioAssayTreatments_assnlist}
\alias{bioAssayTuple}
\alias{bioAssayTuples_assnlist}
\alias{bioDataCube}
\alias{bioDataTuples}
\alias{bioDataValues_assn}
\alias{bioEvent_package}
\alias{biologicalCharacteristics_assnreflist}
\alias{bioMaterial_assnlist}
\alias{bioMaterial_assnref}
\alias{bioMaterialMeasurement}
\alias{bioMaterial_package}
\alias{bioMaterialPlateCol}
\alias{bioMaterialPlateIdentifier}
\alias{bioMaterialPlateRow}
\alias{bioSample}
\alias{bioSample_ref}
\alias{bioSequence}
\alias{bioSequence_assnlist}
\alias{bioSequence_package}
\alias{bioSequence_ref}
\alias{bioSource}
\alias{bioSource_ref}
\alias{bqs_package}
\alias{category_assn}
\alias{channel}
\alias{channel_assnlist}
\alias{channel_assnref}
\alias{channel_ref}
\alias{channels_assnreflist}
\alias{characteristics_assnlist}
\alias{clusterBioAssayData_assnref}
\alias{column}
\alias{componentCompounds_assnlist}
\alias{composite_assnref}
\alias{compositeCompositeMap}
\alias{compositeCompositeMap_assnlist}
\alias{compositeCompositeMap_ref}
\alias{compositeCompositeMaps_assnreflist}
\alias{compositeGroup}
\alias{compositeGroup_assnlist}
\alias{compositeGroup_ref}
\alias{compositeGroups_assnreflist}
\alias{compositePosition}
\alias{compositePositionSources_assnlist}
\alias{compositeSequence}
\alias{compositeSequence_assnlist}
\alias{compositeSequence_assnref}
\alias{compositeSequenceDimension}
\alias{compositeSequenceDimension_ref}
\alias{compositeSequence_ref}
\alias{compositeSequences_assnreflist}
\alias{compound}
\alias{compound_assnlist}
\alias{compound_assnref}
\alias{compoundIndices_assnlist}
\alias{compoundMeasurement}
\alias{compoundMeasurements_assnlist}
\alias{compound_ref}
\alias{concentrationUnit}
\alias{confidenceIndicators_assnreflist}
\alias{contact_assnlist}
\alias{contacts_assnreflist}
\alias{containedFeatures_assnreflist}
\alias{controlFeatures_assnreflist}
\alias{controlledFeatures_assnreflist}
\alias{controlType_assn}
\alias{coordinate_assn}
\alias{database}
\alias{database_assnlist}
\alias{database_assnref}
\alias{databaseEntry}
\alias{database_ref}
\alias{databaseReferences_assnlist}
\alias{dataExternal}
\alias{dataExternal_assn}
\alias{dataFormat}
\alias{dataFormatInfoURI}
\alias{dataInternal}
\alias{dataInternal_assn}
\alias{dataType_assn}
\alias{datum}
\alias{datum_assn}
\alias{defaultValue_assn}
\alias{defectType_assn}
\alias{deltaX}
\alias{deltaY}
\alias{derivedBioAssay}
\alias{derivedBioAssayData}
\alias{derivedBioAssayData_assnreflist}
\alias{derivedBioAssayData_ref}
\alias{derivedBioAssayMap_assnreflist}
\alias{derivedBioAssay_ref}
\alias{derivedSignal}
\alias{derivedSignal_ref}
\alias{description}
\alias{description_package}
\alias{descriptions_assnlist}
\alias{designElement_assnref}
\alias{designElementDimension_assnlist}
\alias{designElementDimension_assnref}
\alias{designElementMapping}
\alias{designElementMapping_assn}
\alias{designElementMaps_assnreflist}
\alias{designElement_package}
\alias{designElementTuple}
\alias{designElementTuples_assnlist}
\alias{designProviders_assnreflist}
\alias{distanceUnit}
\alias{distanceUnit_assn}
\alias{editor}
\alias{email}
\alias{error}
\alias{error_ref}
\alias{expectedValue}
\alias{expectedValue_ref}
\alias{experiment}
\alias{experimentalFactor}
\alias{experimentalFactor_ref}
\alias{experimentalFactors_assnlist}
\alias{experiment_assnlist}
\alias{experimentDesign}
\alias{experimentDesigns_assnlist}
\alias{experiment_package}
\alias{exportedFromDB}
\alias{exportedFromServer}
\alias{exportID}
\alias{exportName}
\alias{externalLIMS_assn}
\alias{externalReference}
\alias{externalReference_assn}
\alias{factorValue}
\alias{factorValue_ref}
\alias{factorValues_assnlist}
\alias{failed}
\alias{failed_ref}
\alias{failTypes_assnlist}
\alias{fax}
\alias{feature}
\alias{feature_assnref}
\alias{featureDefect}
\alias{featureDefects_assnlist}
\alias{featureDimension}
\alias{featureDimension_ref}
\alias{featureExtraction}
\alias{featureExtraction_assn}
\alias{featureGroup}
\alias{featureGroup_ref}
\alias{featureGroups_assnlist}
\alias{featureHeight}
\alias{featureInformation}
\alias{featureInformationSources_assnlist}
\alias{featureLength}
\alias{featureLIMSs_assnlist}
\alias{featureLocation}
\alias{featureLocation_assn}
\alias{feature_ref}
\alias{featureReporterMap}
\alias{featureReporterMap_assnlist}
\alias{featureReporterMap_ref}
\alias{featureReporterMaps_assnreflist}
\alias{features_assnlist}
\alias{featureShape_assn}
\alias{featureWidth}
\alias{fiducial}
\alias{fiducials_assnlist}
\alias{fiducialType_assn}
\alias{filenameURI}
\alias{firstName}
\alias{format_assn}
\alias{hardware}
\alias{hardwareApplication}
\alias{hardwareApplications_assnlist}
\alias{hardware_assnlist}
\alias{hardware_assnref}
\alias{hardwareManufacturers_assnreflist}
\alias{hardware_ref}
\alias{hardwares_assnreflist}
\alias{higherLevelAnalysis_package}
\alias{hybridization}
\alias{identifier}
\alias{identifierLIMS_assn}
\alias{imageAcquisition}
\alias{image_ref}
\alias{images_assnreflist}
\alias{immobilizedCharacteristics_assnreflist}
\alias{information_assnref}
\alias{isApproximateLength}
\alias{isBackground}
\alias{isCircular}
\alias{isSolvent}
\alias{issue}
\alias{kindCV}
\alias{labeledExtract}
\alias{labeledExtract_ref}
\alias{labels_assnreflist}
\alias{lastName}
\alias{lowerRightX}
\alias{lowerRightY}
\alias{make}
\alias{manufactureLIMS}
\alias{manufactureLIMSBiomaterial}
\alias{manufacturingDate}
\alias{massUnit}
\alias{materialType_assn}
\alias{measuredBioAssay}
\alias{measuredBioAssayData}
\alias{measuredBioAssayData_assnreflist}
\alias{measuredBioAssayData_ref}
\alias{measuredBioAssay_ref}
\alias{measuredSignal}
\alias{measuredSignal_ref}
\alias{measurement}
\alias{measurement_assn}
\alias{measurement_package}
\alias{members_assnreflist}
\alias{midInitials}
\alias{mismatchInformation}
\alias{mismatchInformation_assnlist}
\alias{model}
\alias{name}
\alias{nameValueType}
\alias{newSequence}
\alias{node}
\alias{nodeContents}
\alias{nodeContents_assnlist}
\alias{nodes_assnlist}
\alias{nodeValue}
\alias{nodeValue_assnlist}
\alias{normalizationDescription_assn}
\alias{numArrays}
\alias{numberOfFeatures}
\alias{numFeaturesPerCol}
\alias{numFeaturesPerRow}
\alias{ontologyEntries_assnlist}
\alias{ontologyEntry}
\alias{ontologyReference_assn}
\alias{organization}
\alias{organization_ref}
\alias{orientationMark}
\alias{orientationMarkPosition}
\alias{originRelativeTo}
\alias{otherKind}
\alias{owner_assnreflist}
\alias{pages}
\alias{parameter}
\alias{parameter_ref}
\alias{parameters_assnlist}
\alias{parameterType_assnref}
\alias{parameterTypes_assnlist}
\alias{parameterValue}
\alias{parameterValues_assnlist}
\alias{parent_assnref}
\alias{performer_assnref}
\alias{performers_assnreflist}
\alias{person_ref}
\alias{phone}
\alias{physicalArrayDesign}
\alias{physicalArrayDesign_ref}
\alias{physicalBioAssay}
\alias{physicalBioAssayData_assnlist}
\alias{physicalBioAssay_ref}
\alias{physicalBioAssaySource_assnref}
\alias{polymerType_assn}
\alias{position}
\alias{position_assn}
\alias{positionDelta}
\alias{positionDelta_assn}
\alias{presentAbsent}
\alias{presentAbsent_ref}
\alias{producerTransformation_assn}
\alias{propertySets_assnlist}
\alias{protocol}
\alias{protocolApplication}
\alias{protocolApplications_assnlist}
\alias{protocol_assnlist}
\alias{protocol_assnref}
\alias{protocol_package}
\alias{protocol_ref}
\alias{providers_assnreflist}
\alias{publication}
\alias{publisher}
\alias{pvalue}
\alias{pvalue_ref}
\alias{quality}
\alias{qualityControlDescription_assn}
\alias{qualityControlStatistics_assnlist}
\alias{quantitationDimension_assnref}
\alias{quantitationType_assnlist}
\alias{quantitationType_assnref}
\alias{quantitationTypeDimension}
\alias{quantitationTypeDimension_assnlist}
\alias{quantitationTypeDimension_assnref}
\alias{quantitationTypeDimension_ref}
\alias{quantitationTypeMap}
\alias{quantitationTypeMap_assnlist}
\alias{quantitationTypeMapping}
\alias{quantitationTypeMapping_assn}
\alias{quantitationTypeMap_ref}
\alias{quantitationTypeMaps_assnreflist}
\alias{quantitationType_package}
\alias{quantitationTypes_assnreflist}
\alias{quantitationTypeTuple}
\alias{quantitationTypeTuples_assnlist}
\alias{quantityUnit}
\alias{ratio}
\alias{ratio_ref}
\alias{regions_assnlist}
\alias{releaseDate}
\alias{replacedLength}
\alias{replicateDescription_assn}
\alias{reporter}
\alias{reporter_assnlist}
\alias{reporter_assnref}
\alias{reporterCompositeMap}
\alias{reporterCompositeMap_assnlist}
\alias{reporterCompositeMap_ref}
\alias{reporterCompositeMaps_assnreflist}
\alias{reporterDimension}
\alias{reporterDimension_ref}
\alias{reporterGroup}
\alias{reporterGroup_assnlist}
\alias{reporterGroup_ref}
\alias{reporterGroups_assnreflist}
\alias{reporterPosition}
\alias{reporterPositionSources_assnlist}
\alias{reporter_ref}
\alias{reporters_assnreflist}
\alias{roles_assnlist}
\alias{scale_assn}
\alias{security}
\alias{security_assnlist}
\alias{security_assnref}
\alias{securityGroup}
\alias{securityGroup_assnlist}
\alias{securityGroup_ref}
\alias{securityGroups_assnreflist}
\alias{security_ref}
\alias{seqFeature}
\alias{seqFeatureLocation}
\alias{seqFeatures_assnlist}
\alias{sequenceDatabases_assnlist}
\alias{sequencePosition}
\alias{serialNumber}
\alias{software}
\alias{softwareApplication}
\alias{softwareApplications_assnlist}
\alias{software_assnlist}
\alias{software_assnref}
\alias{softwareManufacturers_assnreflist}
\alias{software_ref}
\alias{softwares_assnreflist}
\alias{sourceBioAssays_assnreflist}
\alias{sourceBioMaterialMeasurements_assnlist}
\alias{sourceContact_assnreflist}
\alias{sourcesQuantitationType_assnreflist}
\alias{spacingBetweenCols}
\alias{spacingBetweenRows}
\alias{spacingsBetweenZonesX}
\alias{spacingsBetweenZonesY}
\alias{specializedQuantitationType}
\alias{specializedQuantitationType_ref}
\alias{species_assn}
\alias{startCoord}
\alias{strandType}
\alias{subregions_assnlist}
\alias{substrateType_assn}
\alias{summaryStatistics_assnlist}
\alias{surfaceType_assn}
\alias{target_assnref}
\alias{targetQuantitationType_assnref}
\alias{technologyType_assn}
\alias{temperatureUnit}
\alias{timeUnit}
\alias{tolerance}
\alias{tollFreePhone}
\alias{topLevelBioAssays_assnreflist}
\alias{transformation}
\alias{treatment}
\alias{treatments_assnlist}
\alias{type}
\alias{type_assn}
\alias{types_assnlist}
\alias{unit_assn}
\alias{unitName}
\alias{unitNameCV}
\alias{upperLeftX}
\alias{upperLeftY}
\alias{uri}
\alias{value_assn}
\alias{volume}
\alias{volumeUnit}
\alias{warningType_assn}
\alias{width}
\alias{x}
\alias{y}
\alias{year}
\alias{zone}
\alias{zone_assnref}
\alias{zoneDefect}
\alias{zoneGroup}
\alias{zoneGroups_assnlist}
\alias{zoneLayout}
\alias{zoneLayout_assn}
\alias{zoneLocations_assnlist}
\alias{zone_ref}
\alias{zonesPerX}
\alias{zonesPerY}

% accessor methods
\alias{accession,DatabaseEntry-method}
\alias{accessions_assnlist,BibliographicReference-method}
\alias{accessionVersion,DatabaseEntry-method}
\alias{action,Audit-method}
\alias{action_assn,Treatment-method}
\alias{actionMeasurement_assn,Treatment-method}
\alias{activityDate,ProtocolApplication-method}
\alias{address,Organization-method}
\alias{address,Person-method}
\alias{adjustments_assnlist,ArrayManufactureDeviation-method}
\alias{affiliation_assnref,Person-method}
\alias{affxArray,Array_assnlist-method}
\alias{affxCategory,OntologyEntry-method}
\alias{affxDate,Audit-method}
\alias{affxEnd,CompositePosition-method}
\alias{affxEnd,ReporterPosition-method}
\alias{affxEnd,SequencePosition-method}
\alias{affxImage,PhysicalBioAssayData_assnlist-method}
\alias{affxOrder,BioDataCube-method}
\alias{affxOrder,Treatment-method}
\alias{affxPerson,Contact_assnlist-method}
\alias{affxPerson,Person_assnlist-method}
\alias{affxRow,FeatureLocation-method}
\alias{affxRow,Zone-method}
\alias{affxSequence,BioSequence-method}
\alias{affxStart,CompositePosition-method}
\alias{affxStart,ReporterPosition-method}
\alias{affxStart,SequencePosition-method}
\alias{affxText,Description-method}
\alias{affxText,Protocol-method}
\alias{affxTitle,BibliographicReference-method}
\alias{affxTitle,Protocol-method}
\alias{analysisResults_assnreflist,Experiment-method}
\alias{annotations_assnlist,Description-method}
\alias{annotations_assnlist,ExperimentalFactor-method}
\alias{array_assnlist,Array_package-method}
\alias{array_assnref,BioAssayCreation-method}
\alias{array_assnref,Hybridization-method}
\alias{arrayDesign,ArrayDesign_assnlist-method}
\alias{arrayDesign_assnlist,ArrayDesign_package-method}
\alias{arrayDesign_assnref,Array-method}
\alias{arrayDesign_package,MAGE-ML-method}
\alias{arrayDesign_ref,ArrayDesign_assnref-method}
\alias{arrayGroup,ArrayGroup_assnlist-method}
\alias{arrayGroup_assnlist,Array_package-method}
\alias{arrayGroup_assnref,Array-method}
\alias{arrayGroup_ref,ArrayGroup_assnref-method}
\alias{arrayIdentifier,Array-method}
\alias{arrayManufacture,ArrayManufacture_assnlist-method}
\alias{arrayManufacture_assnlist,Array_package-method}
\alias{arrayManufactureDeviation,ArrayManufactureDeviations_assnlist-method}
\alias{arrayManufactureDeviations_assnlist,Array-method}
\alias{arrayManufacture_ref,Information_assnref-method}
\alias{arrayManufacturers_assnreflist,ArrayManufacture-method}
\alias{array_package,MAGE-ML-method}
\alias{array_ref,Array_assnref-method}
\alias{array_ref,Arrays_assnreflist-method}
\alias{arrays_assnreflist,ArrayGroup-method}
\alias{arrays_assnreflist,ArrayManufacture-method}
\alias{arraySpacingX,ArrayGroup-method}
\alias{arraySpacingY,ArrayGroup-method}
\alias{arrayXOrigin,Array-method}
\alias{arrayYOrigin,Array-method}
\alias{associations_assnlist,OntologyEntry-method}
\alias{audit,AuditTrail_assnlist-method}
\alias{auditAndSecurity_package,MAGE-ML-method}
\alias{auditTrail_assnlist,Array-method}
\alias{auditTrail_assnlist,ArrayDesign-method}
\alias{auditTrail_assnlist,ArrayGroup-method}
\alias{auditTrail_assnlist,ArrayManufacture-method}
\alias{auditTrail_assnlist,Audit-method}
\alias{auditTrail_assnlist,BibliographicReference-method}
\alias{auditTrail_assnlist,BioAssayCreation-method}
\alias{auditTrail_assnlist,BioAssayDataCluster-method}
\alias{auditTrail_assnlist,BioAssayDimension-method}
\alias{auditTrail_assnlist,BioAssayMap-method}
\alias{auditTrail_assnlist,BioAssayTreatment-method}
\alias{auditTrail_assnlist,BioSample-method}
\alias{auditTrail_assnlist,BioSequence-method}
\alias{auditTrail_assnlist,BioSource-method}
\alias{auditTrail_assnlist,Channel-method}
\alias{auditTrail_assnlist,CompositeCompositeMap-method}
\alias{auditTrail_assnlist,CompositeGroup-method}
\alias{auditTrail_assnlist,CompositeSequence-method}
\alias{auditTrail_assnlist,CompositeSequenceDimension-method}
\alias{auditTrail_assnlist,Compound-method}
\alias{auditTrail_assnlist,Database-method}
\alias{auditTrail_assnlist,DerivedBioAssay-method}
\alias{auditTrail_assnlist,DerivedBioAssayData-method}
\alias{auditTrail_assnlist,DerivedSignal-method}
\alias{auditTrail_assnlist,Description-method}
\alias{auditTrail_assnlist,Error-method}
\alias{auditTrail_assnlist,ExpectedValue-method}
\alias{auditTrail_assnlist,Experiment-method}
\alias{auditTrail_assnlist,ExperimentalFactor-method}
\alias{auditTrail_assnlist,ExperimentDesign-method}
\alias{auditTrail_assnlist,FactorValue-method}
\alias{auditTrail_assnlist,Failed-method}
\alias{auditTrail_assnlist,Feature-method}
\alias{auditTrail_assnlist,FeatureDimension-method}
\alias{auditTrail_assnlist,FeatureExtraction-method}
\alias{auditTrail_assnlist,FeatureGroup-method}
\alias{auditTrail_assnlist,FeatureReporterMap-method}
\alias{auditTrail_assnlist,Fiducial-method}
\alias{auditTrail_assnlist,Hardware-method}
\alias{auditTrail_assnlist,HardwareApplication-method}
\alias{auditTrail_assnlist,Hybridization-method}
\alias{auditTrail_assnlist,Image-method}
\alias{auditTrail_assnlist,ImageAcquisition-method}
\alias{auditTrail_assnlist,LabeledExtract-method}
\alias{auditTrail_assnlist,MAGE-ML-method}
\alias{auditTrail_assnlist,ManufactureLIMS-method}
\alias{auditTrail_assnlist,ManufactureLIMSBiomaterial-method}
\alias{auditTrail_assnlist,MeasuredBioAssay-method}
\alias{auditTrail_assnlist,MeasuredBioAssayData-method}
\alias{auditTrail_assnlist,MeasuredSignal-method}
\alias{auditTrail_assnlist,Node-method}
\alias{auditTrail_assnlist,NodeContents-method}
\alias{auditTrail_assnlist,Organization-method}
\alias{auditTrail_assnlist,Parameter-method}
\alias{auditTrail_assnlist,Person-method}
\alias{auditTrail_assnlist,PhysicalArrayDesign-method}
\alias{auditTrail_assnlist,PhysicalBioAssay-method}
\alias{auditTrail_assnlist,PresentAbsent-method}
\alias{auditTrail_assnlist,Protocol-method}
\alias{auditTrail_assnlist,ProtocolApplication-method}
\alias{auditTrail_assnlist,PValue-method}
\alias{auditTrail_assnlist,QuantitationTypeDimension-method}
\alias{auditTrail_assnlist,QuantitationTypeMap-method}
\alias{auditTrail_assnlist,Ratio-method}
\alias{auditTrail_assnlist,Reporter-method}
\alias{auditTrail_assnlist,ReporterCompositeMap-method}
\alias{auditTrail_assnlist,ReporterDimension-method}
\alias{auditTrail_assnlist,ReporterGroup-method}
\alias{auditTrail_assnlist,Security-method}
\alias{auditTrail_assnlist,SecurityGroup-method}
\alias{auditTrail_assnlist,SeqFeature-method}
\alias{auditTrail_assnlist,Software-method}
\alias{auditTrail_assnlist,SoftwareApplication-method}
\alias{auditTrail_assnlist,SpecializedQuantitationType-method}
\alias{auditTrail_assnlist,Transformation-method}
\alias{auditTrail_assnlist,Treatment-method}
\alias{auditTrail_assnlist,Zone-method}
\alias{authors,BibliographicReference-method}
\alias{barcode,ArrayGroup-method}
\alias{basis,SeqFeature-method}
\alias{bibliographicReference,BibliographicReferences_assnlist-method}
\alias{bibliographicReferences_assnlist,Description-method}
\alias{bioAssay_assnlist,BioAssay_package-method}
\alias{bioAssay_assnref,BioAssayTuple-method}
\alias{bioAssayCreation,BioAssayCreation_assn-method}
\alias{bioAssayCreation_assn,PhysicalBioAssay-method}
\alias{bioAssayData_assnlist,BioAssayData_package-method}
\alias{bioAssayData_assnreflist,Experiment-method}
\alias{bioAssayDataCluster,BioAssayDataCluster_assnlist-method}
\alias{bioAssayDataCluster_assnlist,HigherLevelAnalysis_package-method}
\alias{bioAssayDataCluster_ref,AnalysisResults_assnreflist-method}
\alias{bioAssayData_package,MAGE-ML-method}
\alias{bioAssayDataSources_assnreflist,Transformation-method}
\alias{bioAssayDimension,BioAssayDimension_assnlist-method}
\alias{bioAssayDimension_assnlist,BioAssayData_package-method}
\alias{bioAssayDimension_assnref,DerivedBioAssayData-method}
\alias{bioAssayDimension_assnref,MeasuredBioAssayData-method}
\alias{bioAssayDimension_assnref,NodeContents-method}
\alias{bioAssayDimension_ref,BioAssayDimension_assnref-method}
\alias{bioAssayFactorValues_assnreflist,DerivedBioAssay-method}
\alias{bioAssayFactorValues_assnreflist,MeasuredBioAssay-method}
\alias{bioAssayFactorValues_assnreflist,PhysicalBioAssay-method}
\alias{bioAssayMap,BioAssayMap_assnlist-method}
\alias{bioAssayMap_assnlist,BioAssayData_package-method}
\alias{bioAssayMapping,BioAssayMapping_assn-method}
\alias{bioAssayMapping_assn,Transformation-method}
\alias{bioAssayMap_ref,BioAssayMaps_assnreflist-method}
\alias{bioAssayMap_ref,DerivedBioAssayMap_assnreflist-method}
\alias{bioAssayMaps_assnreflist,BioAssayMapping-method}
\alias{bioAssayMapTarget_assnref,BioAssayMap-method}
\alias{bioAssay_package,MAGE-ML-method}
\alias{bioAssays_assnreflist,BioAssayDimension-method}
\alias{bioAssays_assnreflist,Experiment-method}
\alias{bioAssayTreatment,BioAssayTreatments_assnlist-method}
\alias{bioAssayTreatments_assnlist,PhysicalBioAssay-method}
\alias{bioAssayTuple,BioAssayTuples_assnlist-method}
\alias{bioAssayTuples_assnlist,BioDataTuples-method}
\alias{bioDataCube,BioDataValues_assn-method}
\alias{bioDataTuples,BioDataValues_assn-method}
\alias{bioDataValues_assn,DerivedBioAssayData-method}
\alias{bioDataValues_assn,MeasuredBioAssayData-method}
\alias{bioEvent_package,MAGE-ML-method}
\alias{biologicalCharacteristics_assnreflist,CompositeSequence-method}
\alias{bioMaterial_assnlist,BioMaterial_package-method}
\alias{bioMaterial_assnref,BioMaterialMeasurement-method}
\alias{bioMaterial_assnref,ManufactureLIMS-method}
\alias{bioMaterial_assnref,ManufactureLIMSBiomaterial-method}
\alias{bioMaterialMeasurement,SourceBioMaterialMeasurements_assnlist-method}
\alias{bioMaterial_package,MAGE-ML-method}
\alias{bioMaterialPlateCol,ManufactureLIMSBiomaterial-method}
\alias{bioMaterialPlateIdentifier,ManufactureLIMSBiomaterial-method}
\alias{bioMaterialPlateRow,ManufactureLIMSBiomaterial-method}
\alias{bioSample,BioMaterial_assnlist-method}
\alias{bioSample_ref,BioMaterial_assnref-method}
\alias{bioSequence,BioSequence_assnlist-method}
\alias{bioSequence_assnlist,BioSequence_package-method}
\alias{bioSequence_package,MAGE-ML-method}
\alias{bioSequence_ref,BiologicalCharacteristics_assnreflist-method}
\alias{bioSequence_ref,ImmobilizedCharacteristics_assnreflist-method}
\alias{bioSource,BioMaterial_assnlist-method}
\alias{bioSource_ref,BioMaterial_assnref-method}
\alias{bqs_package,MAGE-ML-method}
\alias{category_assn,ExperimentalFactor-method}
\alias{channel,Channel_assnlist-method}
\alias{channel_assnlist,BioAssay_package-method}
\alias{channel_assnref,DerivedSignal-method}
\alias{channel_assnref,Error-method}
\alias{channel_assnref,ExpectedValue-method}
\alias{channel_assnref,Failed-method}
\alias{channel_assnref,MeasuredSignal-method}
\alias{channel_assnref,PresentAbsent-method}
\alias{channel_assnref,PValue-method}
\alias{channel_assnref,Ratio-method}
\alias{channel_assnref,SpecializedQuantitationType-method}
\alias{channel_ref,Channel_assnref-method}
\alias{channel_ref,Channels_assnreflist-method}
\alias{channels_assnreflist,DerivedBioAssay-method}
\alias{channels_assnreflist,Image-method}
\alias{channels_assnreflist,MeasuredBioAssay-method}
\alias{channels_assnreflist,PhysicalBioAssay-method}
\alias{characteristics_assnlist,BioSample-method}
\alias{characteristics_assnlist,BioSource-method}
\alias{characteristics_assnlist,LabeledExtract-method}
\alias{clusterBioAssayData_assnref,BioAssayDataCluster-method}
\alias{column,FeatureLocation-method}
\alias{column,Zone-method}
\alias{componentCompounds_assnlist,Compound-method}
\alias{composite_assnref,CompositePosition-method}
\alias{compositeCompositeMap,CompositeCompositeMap_assnlist-method}
\alias{compositeCompositeMap,DesignElementMap_assnlist-method}
\alias{compositeCompositeMap_assnlist,DesignElement_package-method}
\alias{compositeCompositeMap_ref,CompositeCompositeMaps_assnreflist-method}
\alias{compositeCompositeMap_ref,DesignElementMaps_assnreflist-method}
\alias{compositeCompositeMaps_assnreflist,CompositeSequence-method}
\alias{compositeGroup,CompositeGroup_assnlist-method}
\alias{compositeGroup_assnlist,ArrayDesign_package-method}
\alias{compositeGroup_ref,CompositeGroups_assnreflist-method}
\alias{compositeGroups_assnreflist,ArrayDesign-method}
\alias{compositeGroups_assnreflist,PhysicalArrayDesign-method}
\alias{compositePosition,CompositePositionSources_assnlist-method}
\alias{compositePosition,Coordinate_assn-method}
\alias{compositePositionSources_assnlist,CompositeCompositeMap-method}
\alias{compositeSequence,CompositeSequence_assnlist-method}
\alias{compositeSequence,DesignElement_assnlist-method}
\alias{compositeSequence_assnlist,DesignElement_package-method}
\alias{compositeSequence_assnref,CompositeCompositeMap-method}
\alias{compositeSequence_assnref,ReporterCompositeMap-method}
\alias{compositeSequenceDimension,DesignElementDimension_assnlist-method}
\alias{compositeSequenceDimension_ref,DesignElementDimension_assnref-method}
\alias{compositeSequence_ref,Composite_assnref-method}
\alias{compositeSequence_ref,CompositeSequence_assnref-method}
\alias{compositeSequence_ref,CompositeSequences_assnreflist-method}
\alias{compositeSequence_ref,DesignElement_assnref-method}
\alias{compositeSequences_assnreflist,CompositeGroup-method}
\alias{compositeSequences_assnreflist,CompositeSequenceDimension-method}
\alias{compound,Compound_assnlist-method}
\alias{compound_assnlist,BioMaterial_package-method}
\alias{compound_assnref,CompoundMeasurement-method}
\alias{compoundIndices_assnlist,Compound-method}
\alias{compoundMeasurement,ComponentCompounds_assnlist-method}
\alias{compoundMeasurement,CompoundMeasurements_assnlist-method}
\alias{compoundMeasurements_assnlist,Treatment-method}
\alias{compound_ref,Compound_assnref-method}
\alias{compound_ref,Labels_assnreflist-method}
\alias{concentrationUnit,Unit_assn-method}
\alias{confidenceIndicators_assnreflist,DerivedSignal-method}
\alias{confidenceIndicators_assnreflist,Error-method}
\alias{confidenceIndicators_assnreflist,ExpectedValue-method}
\alias{confidenceIndicators_assnreflist,Failed-method}
\alias{confidenceIndicators_assnreflist,MeasuredSignal-method}
\alias{confidenceIndicators_assnreflist,PresentAbsent-method}
\alias{confidenceIndicators_assnreflist,PValue-method}
\alias{confidenceIndicators_assnreflist,Ratio-method}
\alias{confidenceIndicators_assnreflist,SpecializedQuantitationType-method}
\alias{contact_assnlist,AuditAndSecurity_package-method}
\alias{contacts_assnreflist,Database-method}
\alias{containedFeatures_assnreflist,FeatureDimension-method}
\alias{controlFeatures_assnreflist,Feature-method}
\alias{controlledFeatures_assnreflist,Feature-method}
\alias{controlType_assn,CompositeSequence-method}
\alias{controlType_assn,Feature-method}
\alias{controlType_assn,Reporter-method}
\alias{coordinate_assn,SeqFeatureLocation-method}
\alias{database,Database_assnlist-method}
\alias{database_assnlist,Description_package-method}
\alias{database_assnref,DatabaseEntry-method}
\alias{databaseEntry,Accessions_assnlist-method}
\alias{databaseEntry,DatabaseReferences_assnlist-method}
\alias{databaseEntry,ExternalLIMS_assn-method}
\alias{databaseEntry,IdentifierLIMS_assn-method}
\alias{databaseEntry,OntologyReference_assn-method}
\alias{databaseEntry,SequenceDatabases_assnlist-method}
\alias{database_ref,Database_assnref-method}
\alias{databaseReferences_assnlist,Description-method}
\alias{dataExternal,DataExternal_assn-method}
\alias{dataExternal_assn,BioDataCube-method}
\alias{dataFormat,DataExternal-method}
\alias{dataFormatInfoURI,DataExternal-method}
\alias{dataInternal,DataInternal_assn-method}
\alias{dataInternal_assn,BioDataCube-method}
\alias{dataType_assn,DerivedSignal-method}
\alias{dataType_assn,Error-method}
\alias{dataType_assn,ExpectedValue-method}
\alias{dataType_assn,Failed-method}
\alias{dataType_assn,MeasuredSignal-method}
\alias{dataType_assn,NodeValue-method}
\alias{dataType_assn,Parameter-method}
\alias{dataType_assn,PresentAbsent-method}
\alias{dataType_assn,PValue-method}
\alias{dataType_assn,Ratio-method}
\alias{dataType_assn,SpecializedQuantitationType-method}
\alias{datum,Datum_assn-method}
\alias{datum_assn,QuantitationTypeTuple-method}
\alias{defaultValue_assn,Parameter-method}
\alias{defectType_assn,FeatureDefect-method}
\alias{defectType_assn,ZoneDefect-method}
\alias{deltaX,PositionDelta-method}
\alias{deltaY,PositionDelta-method}
\alias{derivedBioAssay,BioAssay_assnlist-method}
\alias{derivedBioAssay,DerivedBioAssay_assnlist-method}
\alias{derivedBioAssayData,BioAssayData_assnlist-method}
\alias{derivedBioAssayData,DerivedBioAssayData_assnlist-method}
\alias{derivedBioAssayData_assnreflist,DerivedBioAssay-method}
\alias{derivedBioAssayData_ref,BioAssayData_assnreflist-method}
\alias{derivedBioAssayData_ref,BioAssayDataSources_assnreflist-method}
\alias{derivedBioAssayData_ref,ClusterBioAssayData_assnref-method}
\alias{derivedBioAssayData_ref,DerivedBioAssayData_assnreflist-method}
\alias{derivedBioAssayData_ref,DerivedBioAssayDataTarget_assnref-method}
\alias{derivedBioAssayMap_assnreflist,DerivedBioAssay-method}
\alias{derivedBioAssay_ref,BioAssay_assnref-method}
\alias{derivedBioAssay_ref,BioAssayMapTarget_assnref-method}
\alias{derivedBioAssay_ref,BioAssays_assnreflist-method}
\alias{derivedBioAssay_ref,SourceBioAssays_assnreflist-method}
\alias{derivedBioAssay_ref,TopLevelBioAssays_assnreflist-method}
\alias{derivedSignal,QuantitationType_assnlist-method}
\alias{derivedSignal_ref,QuantitationType_assnref-method}
\alias{derivedSignal_ref,QuantitationTypes_assnreflist-method}
\alias{derivedSignal_ref,SourcesQuantitationType_assnreflist-method}
\alias{derivedSignal_ref,TargetQuantitationType_assnref-method}
\alias{description,Descriptions_assnlist-method}
\alias{description,NormalizationDescription_assn-method}
\alias{description,OntologyEntry-method}
\alias{description,QualityControlDescription_assn-method}
\alias{description,ReplicateDescription_assn-method}
\alias{description_package,MAGE-ML-method}
\alias{descriptions_assnlist,Array-method}
\alias{descriptions_assnlist,ArrayDesign-method}
\alias{descriptions_assnlist,ArrayGroup-method}
\alias{descriptions_assnlist,ArrayManufacture-method}
\alias{descriptions_assnlist,Audit-method}
\alias{descriptions_assnlist,BibliographicReference-method}
\alias{descriptions_assnlist,BioAssayCreation-method}
\alias{descriptions_assnlist,BioAssayDataCluster-method}
\alias{descriptions_assnlist,BioAssayDimension-method}
\alias{descriptions_assnlist,BioAssayMap-method}
\alias{descriptions_assnlist,BioAssayTreatment-method}
\alias{descriptions_assnlist,BioSample-method}
\alias{descriptions_assnlist,BioSequence-method}
\alias{descriptions_assnlist,BioSource-method}
\alias{descriptions_assnlist,Channel-method}
\alias{descriptions_assnlist,CompositeCompositeMap-method}
\alias{descriptions_assnlist,CompositeGroup-method}
\alias{descriptions_assnlist,CompositeSequence-method}
\alias{descriptions_assnlist,CompositeSequenceDimension-method}
\alias{descriptions_assnlist,Compound-method}
\alias{descriptions_assnlist,Database-method}
\alias{descriptions_assnlist,DerivedBioAssay-method}
\alias{descriptions_assnlist,DerivedBioAssayData-method}
\alias{descriptions_assnlist,DerivedSignal-method}
\alias{descriptions_assnlist,Description-method}
\alias{descriptions_assnlist,Error-method}
\alias{descriptions_assnlist,ExpectedValue-method}
\alias{descriptions_assnlist,Experiment-method}
\alias{descriptions_assnlist,ExperimentalFactor-method}
\alias{descriptions_assnlist,ExperimentDesign-method}
\alias{descriptions_assnlist,FactorValue-method}
\alias{descriptions_assnlist,Failed-method}
\alias{descriptions_assnlist,Feature-method}
\alias{descriptions_assnlist,FeatureDimension-method}
\alias{descriptions_assnlist,FeatureExtraction-method}
\alias{descriptions_assnlist,FeatureGroup-method}
\alias{descriptions_assnlist,FeatureReporterMap-method}
\alias{descriptions_assnlist,Fiducial-method}
\alias{descriptions_assnlist,Hardware-method}
\alias{descriptions_assnlist,HardwareApplication-method}
\alias{descriptions_assnlist,Hybridization-method}
\alias{descriptions_assnlist,Image-method}
\alias{descriptions_assnlist,ImageAcquisition-method}
\alias{descriptions_assnlist,LabeledExtract-method}
\alias{descriptions_assnlist,MAGE-ML-method}
\alias{descriptions_assnlist,ManufactureLIMS-method}
\alias{descriptions_assnlist,ManufactureLIMSBiomaterial-method}
\alias{descriptions_assnlist,MeasuredBioAssay-method}
\alias{descriptions_assnlist,MeasuredBioAssayData-method}
\alias{descriptions_assnlist,MeasuredSignal-method}
\alias{descriptions_assnlist,Node-method}
\alias{descriptions_assnlist,NodeContents-method}
\alias{descriptions_assnlist,Organization-method}
\alias{descriptions_assnlist,Parameter-method}
\alias{descriptions_assnlist,Person-method}
\alias{descriptions_assnlist,PhysicalArrayDesign-method}
\alias{descriptions_assnlist,PhysicalBioAssay-method}
\alias{descriptions_assnlist,PresentAbsent-method}
\alias{descriptions_assnlist,Protocol-method}
\alias{descriptions_assnlist,ProtocolApplication-method}
\alias{descriptions_assnlist,PValue-method}
\alias{descriptions_assnlist,QuantitationTypeDimension-method}
\alias{descriptions_assnlist,QuantitationTypeMap-method}
\alias{descriptions_assnlist,Ratio-method}
\alias{descriptions_assnlist,Reporter-method}
\alias{descriptions_assnlist,ReporterCompositeMap-method}
\alias{descriptions_assnlist,ReporterDimension-method}
\alias{descriptions_assnlist,ReporterGroup-method}
\alias{descriptions_assnlist,Security-method}
\alias{descriptions_assnlist,SecurityGroup-method}
\alias{descriptions_assnlist,SeqFeature-method}
\alias{descriptions_assnlist,Software-method}
\alias{descriptions_assnlist,SoftwareApplication-method}
\alias{descriptions_assnlist,SpecializedQuantitationType-method}
\alias{descriptions_assnlist,Transformation-method}
\alias{descriptions_assnlist,Treatment-method}
\alias{descriptions_assnlist,Zone-method}
\alias{designElement_assnref,DesignElementTuple-method}
\alias{designElementDimension_assnlist,BioAssayData_package-method}
\alias{designElementDimension_assnref,DerivedBioAssayData-method}
\alias{designElementDimension_assnref,MeasuredBioAssayData-method}
\alias{designElementDimension_assnref,NodeContents-method}
\alias{designElementMapping,DesignElementMapping_assn-method}
\alias{designElementMapping_assn,Transformation-method}
\alias{designElementMaps_assnreflist,DesignElementMapping-method}
\alias{designElement_package,MAGE-ML-method}
\alias{designElementTuple,DesignElementTuples_assnlist-method}
\alias{designElementTuples_assnlist,BioAssayTuple-method}
\alias{designProviders_assnreflist,ArrayDesign-method}
\alias{designProviders_assnreflist,PhysicalArrayDesign-method}
\alias{distanceUnit,DistanceUnit_assn-method}
\alias{distanceUnit,Unit_assn-method}
\alias{distanceUnit_assn,ArrayGroup-method}
\alias{distanceUnit_assn,FeatureGroup-method}
\alias{distanceUnit_assn,Fiducial-method}
\alias{distanceUnit_assn,Position-method}
\alias{distanceUnit_assn,PositionDelta-method}
\alias{distanceUnit_assn,Zone-method}
\alias{distanceUnit_assn,ZoneGroup-method}
\alias{distanceUnit_assn,ZoneLayout-method}
\alias{editor,BibliographicReference-method}
\alias{email,Organization-method}
\alias{email,Person-method}
\alias{error,ConfidenceIndicator_assnlist-method}
\alias{error,QuantitationType_assnlist-method}
\alias{error_ref,ConfidenceIndicators_assnreflist-method}
\alias{error_ref,QuantitationType_assnref-method}
\alias{error_ref,QuantitationTypes_assnreflist-method}
\alias{error_ref,SourcesQuantitationType_assnreflist-method}
\alias{error_ref,TargetQuantitationType_assnref-method}
\alias{expectedValue,ConfidenceIndicator_assnlist-method}
\alias{expectedValue,QuantitationType_assnlist-method}
\alias{expectedValue_ref,ConfidenceIndicators_assnreflist-method}
\alias{expectedValue_ref,QuantitationType_assnref-method}
\alias{expectedValue_ref,QuantitationTypes_assnreflist-method}
\alias{expectedValue_ref,SourcesQuantitationType_assnreflist-method}
\alias{expectedValue_ref,TargetQuantitationType_assnref-method}
\alias{experiment,Experiment_assnlist-method}
\alias{experimentalFactor,ExperimentalFactors_assnlist-method}
\alias{experimentalFactor_ref,ExperimentalFactor_assnref-method}
\alias{experimentalFactors_assnlist,ExperimentDesign-method}
\alias{experiment_assnlist,Experiment_package-method}
\alias{experimentDesign,ExperimentDesigns_assnlist-method}
\alias{experimentDesigns_assnlist,Experiment-method}
\alias{experiment_package,MAGE-ML-method}
\alias{exportedFromDB,ExternalReference-method}
\alias{exportedFromServer,ExternalReference-method}
\alias{exportID,ExternalReference-method}
\alias{exportName,ExternalReference-method}
\alias{externalLIMS_assn,Compound-method}
\alias{externalReference,ExternalReference_assn-method}
\alias{externalReference_assn,Description-method}
\alias{factorValue,FactorValues_assnlist-method}
\alias{factorValue_ref,BioAssayFactorValues_assnreflist-method}
\alias{factorValues_assnlist,ExperimentalFactor-method}
\alias{failed,QuantitationType_assnlist-method}
\alias{failed_ref,QuantitationType_assnref-method}
\alias{failed_ref,QuantitationTypes_assnreflist-method}
\alias{failed_ref,SourcesQuantitationType_assnreflist-method}
\alias{failed_ref,TargetQuantitationType_assnref-method}
\alias{failTypes_assnlist,Reporter-method}
\alias{fax,Organization-method}
\alias{fax,Person-method}
\alias{feature,DesignElement_assnlist-method}
\alias{feature,Features_assnlist-method}
\alias{feature_assnref,FeatureDefect-method}
\alias{feature_assnref,FeatureInformation-method}
\alias{feature_assnref,ManufactureLIMS-method}
\alias{feature_assnref,ManufactureLIMSBiomaterial-method}
\alias{featureDefect,FeatureDefects_assnlist-method}
\alias{featureDefects_assnlist,ArrayManufactureDeviation-method}
\alias{featureDimension,DesignElementDimension_assnlist-method}
\alias{featureDimension_ref,DesignElementDimension_assnref-method}
\alias{featureExtraction,FeatureExtraction_assn-method}
\alias{featureExtraction_assn,MeasuredBioAssay-method}
\alias{featureGroup,FeatureGroups_assnlist-method}
\alias{featureGroup_ref,FeatureGroup_assnref-method}
\alias{featureGroups_assnlist,ArrayDesign-method}
\alias{featureGroups_assnlist,PhysicalArrayDesign-method}
\alias{featureHeight,FeatureGroup-method}
\alias{featureInformation,FeatureInformationSources_assnlist-method}
\alias{featureInformationSources_assnlist,FeatureReporterMap-method}
\alias{featureLength,FeatureGroup-method}
\alias{featureLIMSs_assnlist,ArrayManufacture-method}
\alias{featureLocation,FeatureLocation_assn-method}
\alias{featureLocation_assn,Feature-method}
\alias{feature_ref,ContainedFeatures_assnreflist-method}
\alias{feature_ref,ControlFeatures_assnreflist-method}
\alias{feature_ref,ControlledFeatures_assnreflist-method}
\alias{feature_ref,DesignElement_assnref-method}
\alias{feature_ref,Feature_assnref-method}
\alias{featureReporterMap,DesignElementMap_assnlist-method}
\alias{featureReporterMap,FeatureReporterMap_assnlist-method}
\alias{featureReporterMap_assnlist,DesignElement_package-method}
\alias{featureReporterMap_ref,DesignElementMaps_assnreflist-method}
\alias{featureReporterMap_ref,FeatureReporterMaps_assnreflist-method}
\alias{featureReporterMaps_assnreflist,Reporter-method}
\alias{features_assnlist,FeatureGroup-method}
\alias{featureShape_assn,FeatureGroup-method}
\alias{featureWidth,FeatureGroup-method}
\alias{fiducial,Fiducials_assnlist-method}
\alias{fiducials_assnlist,ArrayGroup-method}
\alias{fiducialType_assn,Fiducial-method}
\alias{filenameURI,DataExternal-method}
\alias{firstName,Person-method}
\alias{format_assn,Image-method}
\alias{hardware,Hardware_assnlist-method}
\alias{hardwareApplication,HardwareApplications_assnlist-method}
\alias{hardwareApplications_assnlist,ProtocolApplication-method}
\alias{hardware_assnlist,Protocol_package-method}
\alias{hardware_assnref,HardwareApplication-method}
\alias{hardware_assnref,Software-method}
\alias{hardwareManufacturers_assnreflist,Hardware-method}
\alias{hardware_ref,Hardware_assnref-method}
\alias{hardware_ref,Hardwares_assnreflist-method}
\alias{hardwares_assnreflist,Protocol-method}
\alias{higherLevelAnalysis_package,MAGE-ML-method}
\alias{hybridization,BioAssayCreation_assn-method}
\alias{identifier,Array-method}
\alias{identifier,ArrayDesign-method}
\alias{identifier,ArrayDesign_ref-method}
\alias{identifier,ArrayGroup-method}
\alias{identifier,ArrayGroup_ref-method}
\alias{identifier,ArrayManufacture-method}
\alias{identifier,ArrayManufacture_ref-method}
\alias{identifier,Array_ref-method}
\alias{identifier,BioAssayCreation-method}
\alias{identifier,BioAssayDataCluster-method}
\alias{identifier,BioAssayDataCluster_ref-method}
\alias{identifier,BioAssayData_ref-method}
\alias{identifier,BioAssayDimension-method}
\alias{identifier,BioAssayDimension_ref-method}
\alias{identifier,BioAssayMap-method}
\alias{identifier,BioAssayMap_ref-method}
\alias{identifier,BioAssay_ref-method}
\alias{identifier,BioAssayTreatment-method}
\alias{identifier,BioMaterial_ref-method}
\alias{identifier,BioSample-method}
\alias{identifier,BioSample_ref-method}
\alias{identifier,BioSequence-method}
\alias{identifier,BioSequence_ref-method}
\alias{identifier,BioSource-method}
\alias{identifier,BioSource_ref-method}
\alias{identifier,Channel-method}
\alias{identifier,Channel_ref-method}
\alias{identifier,CompositeCompositeMap-method}
\alias{identifier,CompositeCompositeMap_ref-method}
\alias{identifier,CompositeGroup-method}
\alias{identifier,CompositeGroup_ref-method}
\alias{identifier,CompositeSequence-method}
\alias{identifier,CompositeSequenceDimension-method}
\alias{identifier,CompositeSequenceDimension_ref-method}
\alias{identifier,CompositeSequence_ref-method}
\alias{identifier,Compound-method}
\alias{identifier,Compound_ref-method}
\alias{identifier,ConfidenceIndicator_ref-method}
\alias{identifier,Contact_ref-method}
\alias{identifier,Database-method}
\alias{identifier,Database_ref-method}
\alias{identifier,DerivedBioAssay-method}
\alias{identifier,DerivedBioAssayData-method}
\alias{identifier,DerivedBioAssayData_ref-method}
\alias{identifier,DerivedBioAssay_ref-method}
\alias{identifier,DerivedSignal-method}
\alias{identifier,DerivedSignal_ref-method}
\alias{identifier,DesignElementDimension_ref-method}
\alias{identifier,DesignElementMap_ref-method}
\alias{identifier,DesignElement_ref-method}
\alias{identifier,Error-method}
\alias{identifier,Error_ref-method}
\alias{identifier,ExpectedValue-method}
\alias{identifier,ExpectedValue_ref-method}
\alias{identifier,Experiment-method}
\alias{identifier,ExperimentalFactor-method}
\alias{identifier,ExperimentalFactor_ref-method}
\alias{identifier,FactorValue-method}
\alias{identifier,FactorValue_ref-method}
\alias{identifier,Failed-method}
\alias{identifier,Failed_ref-method}
\alias{identifier,Feature-method}
\alias{identifier,FeatureDimension-method}
\alias{identifier,FeatureDimension_ref-method}
\alias{identifier,FeatureExtraction-method}
\alias{identifier,FeatureGroup-method}
\alias{identifier,FeatureGroup_ref-method}
\alias{identifier,Feature_ref-method}
\alias{identifier,FeatureReporterMap-method}
\alias{identifier,FeatureReporterMap_ref-method}
\alias{identifier,Hardware-method}
\alias{identifier,Hardware_ref-method}
\alias{identifier,Hybridization-method}
\alias{identifier,Image-method}
\alias{identifier,ImageAcquisition-method}
\alias{identifier,Image_ref-method}
\alias{identifier,LabeledExtract-method}
\alias{identifier,LabeledExtract_ref-method}
\alias{identifier,MAGE-ML-method}
\alias{identifier,MeasuredBioAssay-method}
\alias{identifier,MeasuredBioAssayData-method}
\alias{identifier,MeasuredBioAssayData_ref-method}
\alias{identifier,MeasuredBioAssay_ref-method}
\alias{identifier,MeasuredSignal-method}
\alias{identifier,MeasuredSignal_ref-method}
\alias{identifier,Organization-method}
\alias{identifier,Organization_ref-method}
\alias{identifier,Parameter-method}
\alias{identifier,Parameter_ref-method}
\alias{identifier,Person-method}
\alias{identifier,Person_ref-method}
\alias{identifier,PhysicalArrayDesign-method}
\alias{identifier,PhysicalArrayDesign_ref-method}
\alias{identifier,PhysicalBioAssay-method}
\alias{identifier,PhysicalBioAssay_ref-method}
\alias{identifier,PresentAbsent-method}
\alias{identifier,PresentAbsent_ref-method}
\alias{identifier,Protocol-method}
\alias{identifier,Protocol_ref-method}
\alias{identifier,PValue-method}
\alias{identifier,PValue_ref-method}
\alias{identifier,QuantitationTypeDimension-method}
\alias{identifier,QuantitationTypeDimension_ref-method}
\alias{identifier,QuantitationTypeMap-method}
\alias{identifier,QuantitationTypeMap_ref-method}
\alias{identifier,QuantitationType_ref-method}
\alias{identifier,Ratio-method}
\alias{identifier,Ratio_ref-method}
\alias{identifier,Reporter-method}
\alias{identifier,ReporterCompositeMap-method}
\alias{identifier,ReporterCompositeMap_ref-method}
\alias{identifier,ReporterDimension-method}
\alias{identifier,ReporterDimension_ref-method}
\alias{identifier,ReporterGroup-method}
\alias{identifier,ReporterGroup_ref-method}
\alias{identifier,Reporter_ref-method}
\alias{identifier,Security-method}
\alias{identifier,SecurityGroup-method}
\alias{identifier,SecurityGroup_ref-method}
\alias{identifier,Security_ref-method}
\alias{identifier,Software-method}
\alias{identifier,Software_ref-method}
\alias{identifier,SpecializedQuantitationType-method}
\alias{identifier,SpecializedQuantitationType_ref-method}
\alias{identifier,StandardQuantitationType_ref-method}
\alias{identifier,Transformation-method}
\alias{identifier,Treatment-method}
\alias{identifier,Zone-method}
\alias{identifier,Zone_ref-method}
\alias{identifierLIMS_assn,ManufactureLIMS-method}
\alias{identifierLIMS_assn,ManufactureLIMSBiomaterial-method}
\alias{imageAcquisition,BioAssayTreatments_assnlist-method}
\alias{image_ref,Images_assnreflist-method}
\alias{images_assnreflist,ImageAcquisition-method}
\alias{immobilizedCharacteristics_assnreflist,Reporter-method}
\alias{information_assnref,Array-method}
\alias{isApproximateLength,BioSequence-method}
\alias{isBackground,DerivedSignal-method}
\alias{isBackground,Error-method}
\alias{isBackground,ExpectedValue-method}
\alias{isBackground,Failed-method}
\alias{isBackground,MeasuredSignal-method}
\alias{isBackground,PresentAbsent-method}
\alias{isBackground,PValue-method}
\alias{isBackground,Ratio-method}
\alias{isBackground,SpecializedQuantitationType-method}
\alias{isCircular,BioSequence-method}
\alias{isSolvent,Compound-method}
\alias{issue,BibliographicReference-method}
\alias{kindCV,Measurement-method}
\alias{labeledExtract,BioMaterial_assnlist-method}
\alias{labeledExtract_ref,BioMaterial_assnref-method}
\alias{labels_assnreflist,Channel-method}
\alias{labels_assnreflist,LabeledExtract-method}
\alias{lastName,Person-method}
\alias{length,ArrayGroup-method}
\alias{length,BioSequence-method}
\alias{lowerRightX,Zone-method}
\alias{lowerRightY,Zone-method}
\alias{make,Hardware-method}
\alias{manufactureLIMS,FeatureLIMSs_assnlist-method}
\alias{manufactureLIMSBiomaterial,FeatureLIMSs_assnlist-method}
\alias{manufacturingDate,ArrayManufacture-method}
\alias{massUnit,Unit_assn-method}
\alias{materialType_assn,BioSample-method}
\alias{materialType_assn,BioSource-method}
\alias{materialType_assn,LabeledExtract-method}
\alias{measuredBioAssay,BioAssay_assnlist-method}
\alias{measuredBioAssay,MeasuredBioAssay_assnlist-method}
\alias{measuredBioAssayData,BioAssayData_assnlist-method}
\alias{measuredBioAssayData,MeasuredBioAssayData_assnlist-method}
\alias{measuredBioAssayData_assnreflist,MeasuredBioAssay-method}
\alias{measuredBioAssayData_ref,BioAssayData_assnreflist-method}
\alias{measuredBioAssayData_ref,BioAssayDataSources_assnreflist-method}
\alias{measuredBioAssayData_ref,ClusterBioAssayData_assnref-method}
\alias{measuredBioAssayData_ref,MeasuredBioAssayData_assnreflist-method}
\alias{measuredBioAssay_ref,BioAssay_assnref-method}
\alias{measuredBioAssay_ref,BioAssays_assnreflist-method}
\alias{measuredBioAssay_ref,MeasuredBioAssayTarget_assnref-method}
\alias{measuredBioAssay_ref,SourceBioAssays_assnreflist-method}
\alias{measuredBioAssay_ref,TopLevelBioAssays_assnreflist-method}
\alias{measuredSignal,QuantitationType_assnlist-method}
\alias{measuredSignal_ref,QuantitationType_assnref-method}
\alias{measuredSignal_ref,QuantitationTypes_assnreflist-method}
\alias{measuredSignal_ref,SourcesQuantitationType_assnreflist-method}
\alias{measuredSignal_ref,TargetQuantitationType_assnref-method}
\alias{measurement,ActionMeasurement_assn-method}
\alias{measurement,DefaultValue_assn-method}
\alias{measurement,Measurement_assn-method}
\alias{measurement_assn,BioMaterialMeasurement-method}
\alias{measurement_assn,CompoundMeasurement-method}
\alias{measurement_assn,FactorValue-method}
\alias{measurement_package,MAGE-ML-method}
\alias{members_assnreflist,SecurityGroup-method}
\alias{midInitials,Person-method}
\alias{mismatchInformation,MismatchInformation_assnlist-method}
\alias{mismatchInformation_assnlist,CompositePosition-method}
\alias{mismatchInformation_assnlist,FeatureInformation-method}
\alias{mismatchInformation_assnlist,ReporterPosition-method}
\alias{model,Hardware-method}
\alias{name,ArrayDesign-method}
\alias{name,ArrayDesign_ref-method}
\alias{name,ArrayGroup-method}
\alias{name,ArrayGroup_ref-method}
\alias{name,ArrayManufacture-method}
\alias{name,ArrayManufacture_ref-method}
\alias{name,Array-method}
\alias{name,Array_ref-method}
\alias{name,BioAssayCreation-method}
\alias{name,BioAssayDataCluster-method}
\alias{name,BioAssayDataCluster_ref-method}
\alias{name,BioAssayData_ref-method}
\alias{name,BioAssayDimension-method}
\alias{name,BioAssayDimension_ref-method}
\alias{name,BioAssayMap-method}
\alias{name,BioAssayMap_ref-method}
\alias{name,BioAssay_ref-method}
\alias{name,BioAssayTreatment-method}
\alias{name,BioMaterial_ref-method}
\alias{name,BioSample-method}
\alias{name,BioSample_ref-method}
\alias{name,BioSequence-method}
\alias{name,BioSequence_ref-method}
\alias{name,BioSource-method}
\alias{name,BioSource_ref-method}
\alias{name,Channel-method}
\alias{name,Channel_ref-method}
\alias{name,CompositeCompositeMap-method}
\alias{name,CompositeCompositeMap_ref-method}
\alias{name,CompositeGroup-method}
\alias{name,CompositeGroup_ref-method}
\alias{name,CompositeSequence-method}
\alias{name,CompositeSequenceDimension-method}
\alias{name,CompositeSequenceDimension_ref-method}
\alias{name,CompositeSequence_ref-method}
\alias{name,Compound-method}
\alias{name,Compound_ref-method}
\alias{name,ConfidenceIndicator_ref-method}
\alias{name,Contact_ref-method}
\alias{name,Database-method}
\alias{name,Database_ref-method}
\alias{name,DerivedBioAssay-method}
\alias{name,DerivedBioAssayData-method}
\alias{name,DerivedBioAssayData_ref-method}
\alias{name,DerivedBioAssay_ref-method}
\alias{name,DerivedSignal-method}
\alias{name,DerivedSignal_ref-method}
\alias{name,DesignElementDimension_ref-method}
\alias{name,DesignElementMap_ref-method}
\alias{name,DesignElement_ref-method}
\alias{name,Error-method}
\alias{name,Error_ref-method}
\alias{name,ExpectedValue-method}
\alias{name,ExpectedValue_ref-method}
\alias{name,Experiment-method}
\alias{name,ExperimentalFactor-method}
\alias{name,ExperimentalFactor_ref-method}
\alias{name,FactorValue-method}
\alias{name,FactorValue_ref-method}
\alias{name,Failed-method}
\alias{name,Failed_ref-method}
\alias{name,Feature-method}
\alias{name,FeatureDimension-method}
\alias{name,FeatureDimension_ref-method}
\alias{name,FeatureExtraction-method}
\alias{name,FeatureGroup-method}
\alias{name,FeatureGroup_ref-method}
\alias{name,Feature_ref-method}
\alias{name,FeatureReporterMap-method}
\alias{name,FeatureReporterMap_ref-method}
\alias{name,Hardware-method}
\alias{name,Hardware_ref-method}
\alias{name,Hybridization-method}
\alias{name,Image-method}
\alias{name,ImageAcquisition-method}
\alias{name,Image_ref-method}
\alias{name,LabeledExtract-method}
\alias{name,LabeledExtract_ref-method}
\alias{name,MAGE-ML-method}
\alias{name,MeasuredBioAssay-method}
\alias{name,MeasuredBioAssayData-method}
\alias{name,MeasuredBioAssayData_ref-method}
\alias{name,MeasuredBioAssay_ref-method}
\alias{name,MeasuredSignal-method}
\alias{name,MeasuredSignal_ref-method}
\alias{name,NameValueType-method}
\alias{name,NodeValue-method}
\alias{name,Organization-method}
\alias{name,Organization_ref-method}
\alias{name,Parameter-method}
\alias{name,Parameter_ref-method}
\alias{name,Person-method}
\alias{name,Person_ref-method}
\alias{name,PhysicalArrayDesign-method}
\alias{name,PhysicalArrayDesign_ref-method}
\alias{name,PhysicalBioAssay-method}
\alias{name,PhysicalBioAssay_ref-method}
\alias{name,PresentAbsent-method}
\alias{name,PresentAbsent_ref-method}
\alias{name,Protocol-method}
\alias{name,Protocol_ref-method}
\alias{name,PValue-method}
\alias{name,PValue_ref-method}
\alias{name,QuantitationTypeDimension-method}
\alias{name,QuantitationTypeDimension_ref-method}
\alias{name,QuantitationTypeMap-method}
\alias{name,QuantitationTypeMap_ref-method}
\alias{name,QuantitationType_ref-method}
\alias{name,Ratio-method}
\alias{name,Ratio_ref-method}
\alias{name,Reporter-method}
\alias{name,ReporterCompositeMap-method}
\alias{name,ReporterCompositeMap_ref-method}
\alias{name,ReporterDimension-method}
\alias{name,ReporterDimension_ref-method}
\alias{name,ReporterGroup-method}
\alias{name,ReporterGroup_ref-method}
\alias{name,Reporter_ref-method}
\alias{name,Security-method}
\alias{name,SecurityGroup-method}
\alias{name,SecurityGroup_ref-method}
\alias{name,Security_ref-method}
\alias{name,Software-method}
\alias{name,Software_ref-method}
\alias{name,SpecializedQuantitationType-method}
\alias{name,SpecializedQuantitationType_ref-method}
\alias{name,StandardQuantitationType_ref-method}
\alias{name,Transformation-method}
\alias{name,Treatment-method}
\alias{name,Zone-method}
\alias{name,Zone_ref-method}
\alias{nameValueType,PropertySets_assnlist-method}
\alias{nameValueType,QualityControlStatistics_assnlist-method}
\alias{nameValueType,SummaryStatistics_assnlist-method}
\alias{newSequence,MismatchInformation-method}
\alias{node,Nodes_assnlist-method}
\alias{nodeContents,NodeContents_assnlist-method}
\alias{nodeContents_assnlist,Node-method}
\alias{nodes_assnlist,BioAssayDataCluster-method}
\alias{nodes_assnlist,Node-method}
\alias{nodeValue,NodeValue_assnlist-method}
\alias{nodeValue_assnlist,Node-method}
\alias{normalizationDescription_assn,ExperimentDesign-method}
\alias{numArrays,ArrayGroup-method}
\alias{numberOfFeatures,ArrayDesign-method}
\alias{numberOfFeatures,PhysicalArrayDesign-method}
\alias{numFeaturesPerCol,ZoneLayout-method}
\alias{numFeaturesPerRow,ZoneLayout-method}
\alias{ontologyEntries_assnlist,BioSequence-method}
\alias{ontologyEntry,Action_assn-method}
\alias{ontologyEntry,Annotations_assnlist-method}
\alias{ontologyEntry,Associations_assnlist-method}
\alias{ontologyEntry,Category_assn-method}
\alias{ontologyEntry,Characteristics_assnlist-method}
\alias{ontologyEntry,CompoundIndices_assnlist-method}
\alias{ontologyEntry,ControlType_assn-method}
\alias{ontologyEntry,DataType_assn-method}
\alias{ontologyEntry,DefectType_assn-method}
\alias{ontologyEntry,FailTypes_assnlist-method}
\alias{ontologyEntry,FeatureShape_assn-method}
\alias{ontologyEntry,FiducialType_assn-method}
\alias{ontologyEntry,Format_assn-method}
\alias{ontologyEntry,MaterialType_assn-method}
\alias{ontologyEntry,OntologyEntries_assnlist-method}
\alias{ontologyEntry,Parameters_assnlist-method}
\alias{ontologyEntry,PolymerType_assn-method}
\alias{ontologyEntry,Roles_assnlist-method}
\alias{ontologyEntry,Scale_assn-method}
\alias{ontologyEntry,Species_assn-method}
\alias{ontologyEntry,SubstrateType_assn-method}
\alias{ontologyEntry,SurfaceType_assn-method}
\alias{ontologyEntry,TechnologyType_assn-method}
\alias{ontologyEntry,Type_assn-method}
\alias{ontologyEntry,Types_assnlist-method}
\alias{ontologyEntry,Value_assn-method}
\alias{ontologyEntry,WarningType_assn-method}
\alias{ontologyReference_assn,OntologyEntry-method}
\alias{organization,Contact_assnlist-method}
\alias{organization,Organization_assnlist-method}
\alias{organization_ref,Affiliation_assnref-method}
\alias{organization_ref,ArrayManufacturers_assnreflist-method}
\alias{organization_ref,Contacts_assnreflist-method}
\alias{organization_ref,DesignProviders_assnreflist-method}
\alias{organization_ref,HardwareManufacturers_assnreflist-method}
\alias{organization_ref,Members_assnreflist-method}
\alias{organization_ref,Owner_assnreflist-method}
\alias{organization_ref,Parent_assnref-method}
\alias{organization_ref,Performer_assnref-method}
\alias{organization_ref,Providers_assnreflist-method}
\alias{organization_ref,SoftwareManufacturers_assnreflist-method}
\alias{organization_ref,SourceContact_assnreflist-method}
\alias{orientationMark,ArrayGroup-method}
\alias{orientationMarkPosition,ArrayGroup-method}
\alias{originRelativeTo,Array-method}
\alias{otherKind,Measurement-method}
\alias{owner_assnreflist,Security-method}
\alias{pages,BibliographicReference-method}
\alias{parameter,ParameterTypes_assnlist-method}
\alias{parameter_ref,ParameterType_assnref-method}
\alias{parameters_assnlist,BibliographicReference-method}
\alias{parameterType_assnref,ParameterValue-method}
\alias{parameterTypes_assnlist,Hardware-method}
\alias{parameterTypes_assnlist,Protocol-method}
\alias{parameterTypes_assnlist,Software-method}
\alias{parameterValue,ParameterValues_assnlist-method}
\alias{parameterValues_assnlist,HardwareApplication-method}
\alias{parameterValues_assnlist,ProtocolApplication-method}
\alias{parameterValues_assnlist,SoftwareApplication-method}
\alias{parent_assnref,Organization-method}
\alias{performer_assnref,Audit-method}
\alias{performers_assnreflist,ProtocolApplication-method}
\alias{person_ref,ArrayManufacturers_assnreflist-method}
\alias{person_ref,Contacts_assnreflist-method}
\alias{person_ref,DesignProviders_assnreflist-method}
\alias{person_ref,HardwareManufacturers_assnreflist-method}
\alias{person_ref,Members_assnreflist-method}
\alias{person_ref,Owner_assnreflist-method}
\alias{person_ref,Performer_assnref-method}
\alias{person_ref,Performers_assnreflist-method}
\alias{person_ref,Providers_assnreflist-method}
\alias{person_ref,SoftwareManufacturers_assnreflist-method}
\alias{person_ref,SourceContact_assnreflist-method}
\alias{phone,Organization-method}
\alias{phone,Person-method}
\alias{physicalArrayDesign,ArrayDesign_assnlist-method}
\alias{physicalArrayDesign_ref,ArrayDesign_assnref-method}
\alias{physicalBioAssay,BioAssay_assnlist-method}
\alias{physicalBioAssay,PhysicalBioAssay_assnlist-method}
\alias{physicalBioAssayData_assnlist,PhysicalBioAssay-method}
\alias{physicalBioAssay_ref,BioAssay_assnref-method}
\alias{physicalBioAssay_ref,BioAssays_assnreflist-method}
\alias{physicalBioAssay_ref,PhysicalBioAssay_assnref-method}
\alias{physicalBioAssay_ref,PhysicalBioAssaySource_assnref-method}
\alias{physicalBioAssay_ref,PhysicalBioAssayTarget_assnref-method}
\alias{physicalBioAssay_ref,SourceBioAssays_assnreflist-method}
\alias{physicalBioAssay_ref,Target_assnref-method}
\alias{physicalBioAssay_ref,TopLevelBioAssays_assnreflist-method}
\alias{physicalBioAssaySource_assnref,FeatureExtraction-method}
\alias{polymerType_assn,BioSequence-method}
\alias{position,Position_assn-method}
\alias{position_assn,Feature-method}
\alias{position_assn,Fiducial-method}
\alias{positionDelta,PositionDelta_assn-method}
\alias{positionDelta_assn,FeatureDefect-method}
\alias{positionDelta_assn,ZoneDefect-method}
\alias{presentAbsent,QuantitationType_assnlist-method}
\alias{presentAbsent_ref,QuantitationType_assnref-method}
\alias{presentAbsent_ref,QuantitationTypes_assnreflist-method}
\alias{presentAbsent_ref,SourcesQuantitationType_assnreflist-method}
\alias{presentAbsent_ref,TargetQuantitationType_assnref-method}
\alias{producerTransformation_assn,DerivedBioAssayData-method}
\alias{propertySets_assnlist,Array-method}
\alias{propertySets_assnlist,ArrayDesign-method}
\alias{propertySets_assnlist,ArrayGroup-method}
\alias{propertySets_assnlist,ArrayManufacture-method}
\alias{propertySets_assnlist,ArrayManufactureDeviation-method}
\alias{propertySets_assnlist,Audit-method}
\alias{propertySets_assnlist,BibliographicReference-method}
\alias{propertySets_assnlist,BioAssayCreation-method}
\alias{propertySets_assnlist,BioAssayDataCluster-method}
\alias{propertySets_assnlist,BioAssayDimension-method}
\alias{propertySets_assnlist,BioAssayMap-method}
\alias{propertySets_assnlist,BioAssayMapping-method}
\alias{propertySets_assnlist,BioAssayTreatment-method}
\alias{propertySets_assnlist,BioDataCube-method}
\alias{propertySets_assnlist,BioDataTuples-method}
\alias{propertySets_assnlist,BioMaterialMeasurement-method}
\alias{propertySets_assnlist,BioSample-method}
\alias{propertySets_assnlist,BioSequence-method}
\alias{propertySets_assnlist,BioSource-method}
\alias{propertySets_assnlist,Channel-method}
\alias{propertySets_assnlist,CompositeCompositeMap-method}
\alias{propertySets_assnlist,CompositeGroup-method}
\alias{propertySets_assnlist,CompositePosition-method}
\alias{propertySets_assnlist,CompositeSequence-method}
\alias{propertySets_assnlist,CompositeSequenceDimension-method}
\alias{propertySets_assnlist,Compound-method}
\alias{propertySets_assnlist,CompoundMeasurement-method}
\alias{propertySets_assnlist,ConcentrationUnit-method}
\alias{propertySets_assnlist,Database-method}
\alias{propertySets_assnlist,DatabaseEntry-method}
\alias{propertySets_assnlist,DerivedBioAssay-method}
\alias{propertySets_assnlist,DerivedBioAssayData-method}
\alias{propertySets_assnlist,DerivedSignal-method}
\alias{propertySets_assnlist,Description-method}
\alias{propertySets_assnlist,DesignElementMapping-method}
\alias{propertySets_assnlist,DistanceUnit-method}
\alias{propertySets_assnlist,Error-method}
\alias{propertySets_assnlist,ExpectedValue-method}
\alias{propertySets_assnlist,Experiment-method}
\alias{propertySets_assnlist,ExperimentalFactor-method}
\alias{propertySets_assnlist,ExperimentDesign-method}
\alias{propertySets_assnlist,ExternalReference-method}
\alias{propertySets_assnlist,FactorValue-method}
\alias{propertySets_assnlist,Failed-method}
\alias{propertySets_assnlist,Feature-method}
\alias{propertySets_assnlist,FeatureDefect-method}
\alias{propertySets_assnlist,FeatureDimension-method}
\alias{propertySets_assnlist,FeatureExtraction-method}
\alias{propertySets_assnlist,FeatureGroup-method}
\alias{propertySets_assnlist,FeatureInformation-method}
\alias{propertySets_assnlist,FeatureLocation-method}
\alias{propertySets_assnlist,FeatureReporterMap-method}
\alias{propertySets_assnlist,Fiducial-method}
\alias{propertySets_assnlist,Hardware-method}
\alias{propertySets_assnlist,HardwareApplication-method}
\alias{propertySets_assnlist,Hybridization-method}
\alias{propertySets_assnlist,Image-method}
\alias{propertySets_assnlist,ImageAcquisition-method}
\alias{propertySets_assnlist,LabeledExtract-method}
\alias{propertySets_assnlist,MAGE-ML-method}
\alias{propertySets_assnlist,ManufactureLIMS-method}
\alias{propertySets_assnlist,ManufactureLIMSBiomaterial-method}
\alias{propertySets_assnlist,MassUnit-method}
\alias{propertySets_assnlist,MeasuredBioAssay-method}
\alias{propertySets_assnlist,MeasuredBioAssayData-method}
\alias{propertySets_assnlist,MeasuredSignal-method}
\alias{propertySets_assnlist,Measurement-method}
\alias{propertySets_assnlist,MismatchInformation-method}
\alias{propertySets_assnlist,NameValueType-method}
\alias{propertySets_assnlist,Node-method}
\alias{propertySets_assnlist,NodeContents-method}
\alias{propertySets_assnlist,NodeValue-method}
\alias{propertySets_assnlist,OntologyEntry-method}
\alias{propertySets_assnlist,Organization-method}
\alias{propertySets_assnlist,Parameter-method}
\alias{propertySets_assnlist,ParameterValue-method}
\alias{propertySets_assnlist,Person-method}
\alias{propertySets_assnlist,PhysicalArrayDesign-method}
\alias{propertySets_assnlist,PhysicalBioAssay-method}
\alias{propertySets_assnlist,Position-method}
\alias{propertySets_assnlist,PositionDelta-method}
\alias{propertySets_assnlist,PresentAbsent-method}
\alias{propertySets_assnlist,Protocol-method}
\alias{propertySets_assnlist,ProtocolApplication-method}
\alias{propertySets_assnlist,PValue-method}
\alias{propertySets_assnlist,QuantitationTypeDimension-method}
\alias{propertySets_assnlist,QuantitationTypeMap-method}
\alias{propertySets_assnlist,QuantitationTypeMapping-method}
\alias{propertySets_assnlist,QuantityUnit-method}
\alias{propertySets_assnlist,Ratio-method}
\alias{propertySets_assnlist,Reporter-method}
\alias{propertySets_assnlist,ReporterCompositeMap-method}
\alias{propertySets_assnlist,ReporterDimension-method}
\alias{propertySets_assnlist,ReporterGroup-method}
\alias{propertySets_assnlist,ReporterPosition-method}
\alias{propertySets_assnlist,Security-method}
\alias{propertySets_assnlist,SecurityGroup-method}
\alias{propertySets_assnlist,SeqFeature-method}
\alias{propertySets_assnlist,SeqFeatureLocation-method}
\alias{propertySets_assnlist,SequencePosition-method}
\alias{propertySets_assnlist,Software-method}
\alias{propertySets_assnlist,SoftwareApplication-method}
\alias{propertySets_assnlist,SpecializedQuantitationType-method}
\alias{propertySets_assnlist,TemperatureUnit-method}
\alias{propertySets_assnlist,TimeUnit-method}
\alias{propertySets_assnlist,Transformation-method}
\alias{propertySets_assnlist,Treatment-method}
\alias{propertySets_assnlist,VolumeUnit-method}
\alias{propertySets_assnlist,Zone-method}
\alias{propertySets_assnlist,ZoneDefect-method}
\alias{propertySets_assnlist,ZoneGroup-method}
\alias{propertySets_assnlist,ZoneLayout-method}
\alias{protocol,Protocol_assnlist-method}
\alias{protocolApplication,ProtocolApplications_assnlist-method}
\alias{protocolApplications_assnlist,ArrayDesign-method}
\alias{protocolApplications_assnlist,ArrayManufacture-method}
\alias{protocolApplications_assnlist,BioAssayCreation-method}
\alias{protocolApplications_assnlist,BioAssayMap-method}
\alias{protocolApplications_assnlist,BioAssayTreatment-method}
\alias{protocolApplications_assnlist,CompositeCompositeMap-method}
\alias{protocolApplications_assnlist,FeatureExtraction-method}
\alias{protocolApplications_assnlist,FeatureReporterMap-method}
\alias{protocolApplications_assnlist,Hybridization-method}
\alias{protocolApplications_assnlist,ImageAcquisition-method}
\alias{protocolApplications_assnlist,PhysicalArrayDesign-method}
\alias{protocolApplications_assnlist,QuantitationTypeMap-method}
\alias{protocolApplications_assnlist,ReporterCompositeMap-method}
\alias{protocolApplications_assnlist,Transformation-method}
\alias{protocolApplications_assnlist,Treatment-method}
\alias{protocol_assnlist,Protocol_package-method}
\alias{protocol_assnref,ProtocolApplication-method}
\alias{protocol_package,MAGE-ML-method}
\alias{protocol_ref,Protocol_assnref-method}
\alias{providers_assnreflist,Experiment-method}
\alias{publication,BibliographicReference-method}
\alias{publisher,BibliographicReference-method}
\alias{pvalue,ConfidenceIndicator_assnlist-method}
\alias{pvalue,QuantitationType_assnlist-method}
\alias{pvalue_ref,ConfidenceIndicators_assnreflist-method}
\alias{pvalue_ref,QuantitationType_assnref-method}
\alias{pvalue_ref,QuantitationTypes_assnreflist-method}
\alias{pvalue_ref,SourcesQuantitationType_assnreflist-method}
\alias{pvalue_ref,TargetQuantitationType_assnref-method}
\alias{quality,ManufactureLIMS-method}
\alias{quality,ManufactureLIMSBiomaterial-method}
\alias{qualityControlDescription_assn,ExperimentDesign-method}
\alias{qualityControlStatistics_assnlist,ArrayManufacture-method}
\alias{qualityControlStatistics_assnlist,BioSample-method}
\alias{qualityControlStatistics_assnlist,BioSource-method}
\alias{qualityControlStatistics_assnlist,LabeledExtract-method}
\alias{quantitationDimension_assnref,NodeContents-method}
\alias{quantitationType_assnlist,QuantitationType_package-method}
\alias{quantitationType_assnref,QuantitationTypeTuple-method}
\alias{quantitationTypeDimension,QuantitationTypeDimension_assnlist-method}
\alias{quantitationTypeDimension_assnlist,BioAssayData_package-method}
\alias{quantitationTypeDimension_assnref,DerivedBioAssayData-method}
\alias{quantitationTypeDimension_assnref,MeasuredBioAssayData-method}
\alias{quantitationTypeDimension_ref,QuantitationDimension_assnref-method}
\alias{quantitationTypeDimension_ref,QuantitationTypeDimension_assnref-method}
\alias{quantitationTypeMap,QuantitationTypeMap_assnlist-method}
\alias{quantitationTypeMap_assnlist,BioAssayData_package-method}
\alias{quantitationTypeMapping,QuantitationTypeMapping_assn-method}
\alias{quantitationTypeMapping_assn,Transformation-method}
\alias{quantitationTypeMap_ref,QuantitationTypeMaps_assnreflist-method}
\alias{quantitationTypeMaps_assnreflist,DerivedSignal-method}
\alias{quantitationTypeMaps_assnreflist,Error-method}
\alias{quantitationTypeMaps_assnreflist,ExpectedValue-method}
\alias{quantitationTypeMaps_assnreflist,Failed-method}
\alias{quantitationTypeMaps_assnreflist,MeasuredSignal-method}
\alias{quantitationTypeMaps_assnreflist,PresentAbsent-method}
\alias{quantitationTypeMaps_assnreflist,PValue-method}
\alias{quantitationTypeMaps_assnreflist,QuantitationTypeMapping-method}
\alias{quantitationTypeMaps_assnreflist,Ratio-method}
\alias{quantitationTypeMaps_assnreflist,SpecializedQuantitationType-method}
\alias{quantitationType_package,MAGE-ML-method}
\alias{quantitationTypes_assnreflist,QuantitationTypeDimension-method}
\alias{quantitationTypeTuple,QuantitationTypeTuples_assnlist-method}
\alias{quantitationTypeTuples_assnlist,DesignElementTuple-method}
\alias{quantityUnit,Unit_assn-method}
\alias{ratio,QuantitationType_assnlist-method}
\alias{ratio_ref,QuantitationType_assnref-method}
\alias{ratio_ref,QuantitationTypes_assnreflist-method}
\alias{ratio_ref,SourcesQuantitationType_assnreflist-method}
\alias{ratio_ref,TargetQuantitationType_assnref-method}
\alias{regions_assnlist,SeqFeature-method}
\alias{releaseDate,SoftwareApplication-method}
\alias{replacedLength,MismatchInformation-method}
\alias{replicateDescription_assn,ExperimentDesign-method}
\alias{reporter,DesignElement_assnlist-method}
\alias{reporter,Reporter_assnlist-method}
\alias{reporter_assnlist,DesignElement_package-method}
\alias{reporter_assnref,FeatureReporterMap-method}
\alias{reporter_assnref,ReporterPosition-method}
\alias{reporterCompositeMap,DesignElementMap_assnlist-method}
\alias{reporterCompositeMap,ReporterCompositeMap_assnlist-method}
\alias{reporterCompositeMap_assnlist,DesignElement_package-method}
\alias{reporterCompositeMap_ref,DesignElementMaps_assnreflist-method}
\alias{reporterCompositeMap_ref,ReporterCompositeMaps_assnreflist-method}
\alias{reporterCompositeMaps_assnreflist,CompositeSequence-method}
\alias{reporterDimension,DesignElementDimension_assnlist-method}
\alias{reporterDimension_ref,DesignElementDimension_assnref-method}
\alias{reporterGroup,ReporterGroup_assnlist-method}
\alias{reporterGroup_assnlist,ArrayDesign_package-method}
\alias{reporterGroup_ref,ReporterGroups_assnreflist-method}
\alias{reporterGroups_assnreflist,ArrayDesign-method}
\alias{reporterGroups_assnreflist,PhysicalArrayDesign-method}
\alias{reporterPosition,Coordinate_assn-method}
\alias{reporterPosition,ReporterPositionSources_assnlist-method}
\alias{reporterPositionSources_assnlist,ReporterCompositeMap-method}
\alias{reporter_ref,DesignElement_assnref-method}
\alias{reporter_ref,Reporter_assnref-method}
\alias{reporter_ref,Reporters_assnreflist-method}
\alias{reporters_assnreflist,ReporterDimension-method}
\alias{reporters_assnreflist,ReporterGroup-method}
\alias{roles_assnlist,Organization-method}
\alias{roles_assnlist,Person-method}
\alias{scale_assn,DerivedSignal-method}
\alias{scale_assn,Error-method}
\alias{scale_assn,ExpectedValue-method}
\alias{scale_assn,Failed-method}
\alias{scale_assn,MeasuredSignal-method}
\alias{scale_assn,NodeValue-method}
\alias{scale_assn,PresentAbsent-method}
\alias{scale_assn,PValue-method}
\alias{scale_assn,Ratio-method}
\alias{scale_assn,SpecializedQuantitationType-method}
\alias{security,Security_assnlist-method}
\alias{security_assnlist,AuditAndSecurity_package-method}
\alias{security_assnref,Array-method}
\alias{security_assnref,ArrayDesign-method}
\alias{security_assnref,ArrayGroup-method}
\alias{security_assnref,ArrayManufacture-method}
\alias{security_assnref,Audit-method}
\alias{security_assnref,BibliographicReference-method}
\alias{security_assnref,BioAssayCreation-method}
\alias{security_assnref,BioAssayDataCluster-method}
\alias{security_assnref,BioAssayDimension-method}
\alias{security_assnref,BioAssayMap-method}
\alias{security_assnref,BioAssayTreatment-method}
\alias{security_assnref,BioSample-method}
\alias{security_assnref,BioSequence-method}
\alias{security_assnref,BioSource-method}
\alias{security_assnref,Channel-method}
\alias{security_assnref,CompositeCompositeMap-method}
\alias{security_assnref,CompositeGroup-method}
\alias{security_assnref,CompositeSequence-method}
\alias{security_assnref,CompositeSequenceDimension-method}
\alias{security_assnref,Compound-method}
\alias{security_assnref,Database-method}
\alias{security_assnref,DerivedBioAssay-method}
\alias{security_assnref,DerivedBioAssayData-method}
\alias{security_assnref,DerivedSignal-method}
\alias{security_assnref,Description-method}
\alias{security_assnref,Error-method}
\alias{security_assnref,ExpectedValue-method}
\alias{security_assnref,Experiment-method}
\alias{security_assnref,ExperimentalFactor-method}
\alias{security_assnref,ExperimentDesign-method}
\alias{security_assnref,FactorValue-method}
\alias{security_assnref,Failed-method}
\alias{security_assnref,Feature-method}
\alias{security_assnref,FeatureDimension-method}
\alias{security_assnref,FeatureExtraction-method}
\alias{security_assnref,FeatureGroup-method}
\alias{security_assnref,FeatureReporterMap-method}
\alias{security_assnref,Fiducial-method}
\alias{security_assnref,Hardware-method}
\alias{security_assnref,HardwareApplication-method}
\alias{security_assnref,Hybridization-method}
\alias{security_assnref,Image-method}
\alias{security_assnref,ImageAcquisition-method}
\alias{security_assnref,LabeledExtract-method}
\alias{security_assnref,MAGE-ML-method}
\alias{security_assnref,ManufactureLIMS-method}
\alias{security_assnref,ManufactureLIMSBiomaterial-method}
\alias{security_assnref,MeasuredBioAssay-method}
\alias{security_assnref,MeasuredBioAssayData-method}
\alias{security_assnref,MeasuredSignal-method}
\alias{security_assnref,Node-method}
\alias{security_assnref,NodeContents-method}
\alias{security_assnref,Organization-method}
\alias{security_assnref,Parameter-method}
\alias{security_assnref,Person-method}
\alias{security_assnref,PhysicalArrayDesign-method}
\alias{security_assnref,PhysicalBioAssay-method}
\alias{security_assnref,PresentAbsent-method}
\alias{security_assnref,Protocol-method}
\alias{security_assnref,ProtocolApplication-method}
\alias{security_assnref,PValue-method}
\alias{security_assnref,QuantitationTypeDimension-method}
\alias{security_assnref,QuantitationTypeMap-method}
\alias{security_assnref,Ratio-method}
\alias{security_assnref,Reporter-method}
\alias{security_assnref,ReporterCompositeMap-method}
\alias{security_assnref,ReporterDimension-method}
\alias{security_assnref,ReporterGroup-method}
\alias{security_assnref,Security-method}
\alias{security_assnref,SecurityGroup-method}
\alias{security_assnref,SeqFeature-method}
\alias{security_assnref,Software-method}
\alias{security_assnref,SoftwareApplication-method}
\alias{security_assnref,SpecializedQuantitationType-method}
\alias{security_assnref,Transformation-method}
\alias{security_assnref,Treatment-method}
\alias{security_assnref,Zone-method}
\alias{securityGroup,SecurityGroup_assnlist-method}
\alias{securityGroup_assnlist,AuditAndSecurity_package-method}
\alias{securityGroup_ref,SecurityGroups_assnreflist-method}
\alias{securityGroups_assnreflist,Security-method}
\alias{security_ref,Security_assnref-method}
\alias{seqFeature,SeqFeatures_assnlist-method}
\alias{seqFeatureLocation,Regions_assnlist-method}
\alias{seqFeatureLocation,Subregions_assnlist-method}
\alias{seqFeatures_assnlist,BioSequence-method}
\alias{sequenceDatabases_assnlist,BioSequence-method}
\alias{sequencePosition,Coordinate_assn-method}
\alias{serialNumber,HardwareApplication-method}
\alias{software,Software_assnlist-method}
\alias{softwareApplication,SoftwareApplications_assnlist-method}
\alias{softwareApplications_assnlist,ProtocolApplication-method}
\alias{software_assnlist,Protocol_package-method}
\alias{software_assnref,SoftwareApplication-method}
\alias{softwareManufacturers_assnreflist,Software-method}
\alias{software_ref,Software_assnref-method}
\alias{software_ref,Softwares_assnreflist-method}
\alias{softwares_assnreflist,Hardware-method}
\alias{softwares_assnreflist,Protocol-method}
\alias{softwares_assnreflist,Software-method}
\alias{sourceBioAssays_assnreflist,BioAssayMap-method}
\alias{sourceBioMaterialMeasurements_assnlist,BioAssayCreation-method}
\alias{sourceBioMaterialMeasurements_assnlist,Hybridization-method}
\alias{sourceBioMaterialMeasurements_assnlist,Treatment-method}
\alias{sourceContact_assnreflist,BioSource-method}
\alias{sourcesQuantitationType_assnreflist,QuantitationTypeMap-method}
\alias{spacingBetweenCols,ZoneLayout-method}
\alias{spacingBetweenRows,ZoneLayout-method}
\alias{spacingsBetweenZonesX,ZoneGroup-method}
\alias{spacingsBetweenZonesY,ZoneGroup-method}
\alias{specializedQuantitationType,QuantitationType_assnlist-method}
\alias{specializedQuantitationType_ref,QuantitationType_assnref-method}
\alias{specializedQuantitationType_ref,QuantitationTypes_assnreflist-method}
\alias{specializedQuantitationType_ref,SourcesQuantitationType_assnreflist-method}
\alias{specializedQuantitationType_ref,TargetQuantitationType_assnref-method}
\alias{species_assn,BioSequence-method}
\alias{species_assn,CompositeGroup-method}
\alias{species_assn,FeatureGroup-method}
\alias{species_assn,ReporterGroup-method}
\alias{startCoord,MismatchInformation-method}
\alias{strandType,SeqFeatureLocation-method}
\alias{subregions_assnlist,SeqFeatureLocation-method}
\alias{substrateType_assn,ArrayGroup-method}
\alias{summaryStatistics_assnlist,DerivedBioAssayData-method}
\alias{summaryStatistics_assnlist,MeasuredBioAssayData-method}
\alias{surfaceType_assn,PhysicalArrayDesign-method}
\alias{target_assnref,BioAssayTreatment-method}
\alias{target_assnref,ImageAcquisition-method}
\alias{targetQuantitationType_assnref,Error-method}
\alias{targetQuantitationType_assnref,ExpectedValue-method}
\alias{targetQuantitationType_assnref,PValue-method}
\alias{targetQuantitationType_assnref,QuantitationTypeMap-method}
\alias{technologyType_assn,FeatureGroup-method}
\alias{temperatureUnit,Unit_assn-method}
\alias{timeUnit,Unit_assn-method}
\alias{tolerance,ArrayManufacture-method}
\alias{tollFreePhone,Organization-method}
\alias{tollFreePhone,Person-method}
\alias{topLevelBioAssays_assnreflist,ExperimentDesign-method}
\alias{transformation,ProducerTransformation_assn-method}
\alias{treatment,Treatments_assnlist-method}
\alias{treatments_assnlist,BioSample-method}
\alias{treatments_assnlist,BioSource-method}
\alias{treatments_assnlist,LabeledExtract-method}
\alias{type,Measurement-method}
\alias{type,NameValueType-method}
\alias{type_assn,BioSample-method}
\alias{type_assn,BioSequence-method}
\alias{type_assn,DatabaseEntry-method}
\alias{type_assn,DerivedBioAssay-method}
\alias{type_assn,Hardware-method}
\alias{type_assn,NodeValue-method}
\alias{type_assn,Protocol-method}
\alias{type_assn,Software-method}
\alias{types_assnlist,CompositeGroup-method}
\alias{types_assnlist,ExperimentDesign-method}
\alias{types_assnlist,FeatureGroup-method}
\alias{types_assnlist,ReporterGroup-method}
\alias{unit_assn,Measurement-method}
\alias{unitName,ConcentrationUnit-method}
\alias{unitName,DistanceUnit-method}
\alias{unitName,MassUnit-method}
\alias{unitName,QuantityUnit-method}
\alias{unitName,TemperatureUnit-method}
\alias{unitName,TimeUnit-method}
\alias{unitName,VolumeUnit-method}
\alias{unitNameCV,ConcentrationUnit-method}
\alias{unitNameCV,DistanceUnit-method}
\alias{unitNameCV,MassUnit-method}
\alias{unitNameCV,QuantityUnit-method}
\alias{unitNameCV,TemperatureUnit-method}
\alias{unitNameCV,TimeUnit-method}
\alias{unitNameCV,VolumeUnit-method}
\alias{upperLeftX,Zone-method}
\alias{upperLeftY,Zone-method}
\alias{uri,BibliographicReference-method}
\alias{uri,Database-method}
\alias{uri,DatabaseEntry-method}
\alias{uri,Description-method}
\alias{uri,Hardware-method}
\alias{uri,Image-method}
\alias{uri,Organization-method}
\alias{uri,Person-method}
\alias{uri,Protocol-method}
\alias{uri,Software-method}
\alias{value,Datum-method}
\alias{value,Measurement-method}
\alias{value,NameValueType-method}
\alias{value,NodeValue-method}
\alias{value,OntologyEntry-method}
\alias{value,ParameterValue-method}
\alias{value_assn,FactorValue-method}
\alias{version,ArrayDesign-method}
\alias{version,Database-method}
\alias{version,PhysicalArrayDesign-method}
\alias{version,SoftwareApplication-method}
\alias{volume,BibliographicReference-method}
\alias{volumeUnit,Unit_assn-method}
\alias{warningType_assn,Reporter-method}
\alias{width,ArrayGroup-method}
\alias{x,Position-method}
\alias{y,Position-method}
\alias{year,BibliographicReference-method}
\alias{zone,ZoneLocations_assnlist-method}
\alias{zone_assnref,Feature-method}
\alias{zone_assnref,ZoneDefect-method}
\alias{zoneDefect,Adjustments_assnlist-method}
\alias{zoneGroup,ZoneGroups_assnlist-method}
\alias{zoneGroups_assnlist,PhysicalArrayDesign-method}
\alias{zoneLayout,ZoneLayout_assn-method}
\alias{zoneLayout_assn,ZoneGroup-method}
\alias{zoneLocations_assnlist,ZoneGroup-method}
\alias{zone_ref,Zone_assnref-method}
\alias{zonesPerX,ZoneGroup-method}
\alias{zonesPerY,ZoneGroup-method}


\title{

  (Auto-generated) classes and methods for navigating DTT sample
  attributes.

}
\description{

  These classes and methods facilitate navigation of the Affymetrix DTT
  sample attribute files. DTT files are generated by the Affymetrix GCOS
  application. Specifically, the classes are R representations of the
  Affymetrix classes defined in MAGE-ML.dtd. The 'show' methods for
  these classs indicate how objects are to be accessed; details are in
  the `Retrieving MAGE and ARR sample attributes' vignette.

}
\section{Objects from the Class}{

  Objects are created internally, by parsing XML retrieved from ARR, and
  are not created directly by the user.

}
\section{Slots}{

  Objects have a variety of slots. To discover slot names and content
  type, see the \code{slotNames} portion of the example, below.

}
\section{Extends}{
Class \code{"\linkS4class{DTTCompatibility}"}, directly.
Class \code{"\linkS4class{AffyCompatible}"}, by class "DTTCompatibility", distance 2.
}
\section{Methods}{

  'Accessor' methods are defined on auto-generated classes to access slot
  content.

  Accessors are apparent from displaying the object; each line of the
  display is of the form \code{accessor: content}. To retrieve the
  content of an instance \code{object}, invoke \code{accessor(object)}.

  Some objects contains several instances of another object. These
  'typed associations' are represented by the
  \code{\linkS4class{.TypedAssociation}} class, and are navigated or
  subset as one would a standard R vector, e.g., using \code{[},
  \code{[[}. A useful paradigm for displaying all instances is
  illustrated in the vignette.

}
\author{Martin Morgan <mtmorgan@fhcrc.org>}
\seealso{

  Vignette `Retrieving MAGE and ARR sample attributes'.

}

\examples{
## defined and auto-generated classes
names(slot(getClass("DTTCompatibility",
                    where=getNamespace("AffyCompatible")),
           "subclasses"))


}

\keyword{classes}
